% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange-SpatVector.R
\encoding{UTF-8}
\name{arrange.SpatVector}
\alias{arrange.SpatVector}
\title{Order a \code{\link[terra:vect]{SpatVector}} using column values}
\usage{
\method{arrange}{SpatVector}(.data, ..., .by_group = FALSE, .locale = NULL)
}
\arguments{
\item{.data}{A \code{SpatVector} created with \code{\link[terra:vect]{terra::vect()}}.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Variables, or
functions of variables. Use \code{\link[dplyr:desc]{desc()}} to sort a variable in descending
order.}

\item{.by_group}{If \code{TRUE}, sort first by grouping variable. This applies to
grouped \code{\link[terra:vect]{SpatVector}} objects only.}

\item{.locale}{The locale to sort character vectors in.
\itemize{
\item If \code{NULL}, the default, uses the \code{"C"} locale unless the deprecated
\code{dplyr.legacy_locale} global option escape hatch is active. See the
\link[dplyr]{dplyr-locale} help page for more details.
\item If a single string from \code{\link[stringi:stri_locale_list]{stringi::stri_locale_list()}} is supplied, then
this will be used as the locale to sort with. For example, \code{"en"} will
sort with the American English locale. This requires the stringi package.
\item If \code{"C"} is supplied, then character vectors will always be sorted in the
C locale. This does not require stringi and is often much faster than
supplying a locale identifier.
}

The C locale is not the same as English locales, such as \code{"en"},
particularly when it comes to data containing a mix of upper and lower case
letters. This is explained in more detail on the \link[dplyr:dplyr-locale]{locale}
help page under the \verb{Default locale} section.}
}
\value{
A \code{\link[terra:vect]{SpatVector}} object.
}
\description{
\code{arrange.SpatVector()} orders the geometries of a \code{SpatVector} by the
values of selected columns.
}
\section{\CRANpkg{terra} equivalent}{


\code{\link[terra:sort]{terra::sort()}}.
}

\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:arrange]{dplyr::arrange()}} method for
\code{\link[terra:vect]{SpatVector}} objects.
}

\examples{

library(terra)
library(dplyr)

v <- vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))

# Single variable

v |>
  arrange(desc(iso2))

# Two variables
v |>
  mutate(even = as.double(cpro) \%\% 2 == 0) |>
  arrange(desc(even), desc(iso2))

# With new variables
v |>
  mutate(area_geom = terra::expanse(v)) |>
  arrange(area_geom)
}
\seealso{
\code{\link[dplyr:arrange]{dplyr::arrange()}}.

Other \CRANpkg{dplyr} verbs that operate on rows:
\code{\link[=distinct.SpatVector]{distinct.SpatVector()}},
\code{\link{filter.Spat}},
\code{\link{rows.SpatVector}},
\code{\link{slice.Spat}}
}
\concept{dplyr.rows}
