% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-coordinates.R
\encoding{UTF-8}
\name{as_coordinates}
\alias{as_coordinates}
\title{Get cell number, row and column from a \code{\link[terra:rast]{SpatRaster}}}
\usage{
as_coordinates(x, as.raster = FALSE)
}
\arguments{
\item{x}{A \code{SpatRaster} object.}

\item{as.raster}{If \code{TRUE}, the result is a \code{SpatRaster} object with three
layers indicating the position of each cell (cell number, row and column).}
}
\value{
A \link[tibble:tbl_df]{tibble} or a \code{\link[terra:rast]{SpatRaster}} (if \code{as.raster = TRUE}) with
one row (or cell) for each cell in \code{x}.

When \code{as.raster = TRUE}, the resulting \code{\link[terra:rast]{SpatRaster}} has the
same CRS,
extent and resolution as \code{x}.
}
\description{
\code{as_coordinates()} returns the position of each cell in the
\code{\link[terra:rast]{SpatRaster}}
matrix.
}
\examples{

library(terra)

f <- system.file("extdata/cyl_temp.tif", package = "tidyterra")

r <- rast(f)

as_coordinates(r)
as_coordinates(r, as.raster = TRUE)

as_coordinates(r, as.raster = TRUE) |> plot()

}
\seealso{
\code{\link[=slice.SpatRaster]{slice.SpatRaster()}}.

Coercing objects:
\code{\link[=as_sf]{as_sf()}},
\code{\link[=as_spatraster]{as_spatraster()}},
\code{\link[=as_spatvector]{as_spatvector()}},
\code{\link{as_tibble.Spat}},
\code{\link{fortify.Spat}},
\code{\link{tidy.Spat}}
}
\concept{coerce}
