% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-sf.R
\encoding{UTF-8}
\name{as_sf}
\alias{as_sf}
\title{Coerce a \code{\link[terra:vect]{SpatVector}} to an \code{\link[sf:st_sf]{sf}} object}
\usage{
as_sf(x, ...)
}
\arguments{
\item{x}{A \code{SpatVector} created with \code{\link[terra:vect]{terra::vect()}}.}

\item{...}{Additional arguments passed to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}.}
}
\value{
An \code{\link[sf:st_sf]{sf}} object with an additional \code{tbl_df} class for
pretty printing.
}
\description{
\code{\link[=as_sf]{as_sf()}} coerces a \code{\link[terra:vect]{SpatVector}} into an \code{sf} object. It wraps
\code{\link[sf:st_as_sf]{sf::st_as_sf()}} and preserves groups created with
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}}.
}
\examples{

library(terra)

f <- system.file("extdata/cyl.gpkg", package = "tidyterra")
v <- terra::vect(f)

# The input is ungrouped.
v
is_grouped_spatvector(v)

# Get ungrouped data.
a_sf <- as_sf(v)

dplyr::is_grouped_df(a_sf)

# Grouped

v$gr <- c("C", "A", "A", "B", "A", "B", "B")
v$gr2 <- rep(c("F", "G", "F"), 3)

gr_v <- group_by(v, gr, gr2)

gr_v
is_grouped_spatvector(gr_v)

group_data(gr_v)

# An sf object.

a_gr_sf <- as_sf(gr_v)

dplyr::is_grouped_df(a_gr_sf)

group_data(a_gr_sf)

}
\seealso{
Coercing objects:
\code{\link[=as_coordinates]{as_coordinates()}},
\code{\link[=as_spatraster]{as_spatraster()}},
\code{\link[=as_spatvector]{as_spatvector()}},
\code{\link{as_tibble.Spat}},
\code{\link{fortify.Spat}},
\code{\link{tidy.Spat}}
}
\concept{coerce}
