% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-spatvector.R
\encoding{UTF-8}
\name{as_spatvector}
\alias{as_spatvector}
\alias{as_spatvector.data.frame}
\alias{as_spatvector.sf}
\alias{as_spatvector.sfc}
\alias{as_spatvector.SpatVector}
\title{Coerce objects to \code{\link[terra:vect]{SpatVector}}}
\usage{
as_spatvector(x, ...)

\method{as_spatvector}{data.frame}(x, ..., geom = c("lon", "lat"), crs = "")

\method{as_spatvector}{sf}(x, ...)

\method{as_spatvector}{sfc}(x, ...)

\method{as_spatvector}{SpatVector}(x, ...)
}
\arguments{
\item{x}{A \link[tibble:tbl_df]{tibble}, data frame or \CRANpkg{sf} object of
class \code{sf} or \code{\link[sf:st_sfc]{sfc}}.}

\item{...}{Additional arguments passed to \code{\link[terra:vect]{terra::vect()}}.}

\item{geom}{Character vector naming the fields that contain the geometry
data. Use two names for point coordinates (\code{x} and \code{y}) or one name for a
column with WKT geometries.}

\item{crs}{A CRS in several formats (PROJ.4, WKT, EPSG code, etc.) or a
spatial object from \CRANpkg{sf} or \CRANpkg{terra} that
includes the target coordinate reference system. See \code{\link[=pull_crs]{pull_crs()}} and
\strong{Details}.}
}
\value{
A \code{\link[terra:vect]{SpatVector}} object.
}
\description{
\code{as_spatvector()} turns an existing object into a
\code{\link[terra:vect]{SpatVector}}. It wraps
the \code{\link[terra:vect]{terra::vect()}} S4 method for the \code{data.frame} signature.
}
\details{
This function differs from \code{\link[terra:vect]{terra::vect()}} in the following ways:
\itemize{
\item Rows with geometry values \code{NA} or \code{""} are removed before conversion.
\item If \code{x} is a grouped data frame (see \code{\link[dplyr:group_by]{dplyr::group_by()}}), the grouping
variables are transferred and a grouped \code{\link[terra:vect]{SpatVector}} is
created (see
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}}).
\item If no \code{crs} is provided and the \link[tibble:tbl_df]{tibble} was created with
\code{\link[=as_tibble.SpatVector]{as_tibble.SpatVector()}}, the \code{crs} is inferred from
\code{\link[=attr]{attr(x, "crs")}}.
\item It handles the conversion of \code{EMPTY} geometries between
\CRANpkg{sf} and \CRANpkg{terra}.
}
}
\section{\CRANpkg{terra} equivalent}{


\code{\link[terra:vect]{terra::vect()}}.
}

\examples{
library(terra)

v <- vect(matrix(1:80, ncol = 2), crs = "EPSG:3857")

v$cat <- sample(LETTERS[1:4], size = nrow(v), replace = TRUE)

v

# Create tibble
as_tbl <- as_tibble(v, geom = "WKT")

as_tbl

# From tibble
newvect <- as_spatvector(as_tbl, geom = "geometry", crs = "EPSG:3857")
newvect

}
\seealso{
\code{\link[=pull_crs]{pull_crs()}} for retrieving CRS and the corresponding utilities
\code{\link[sf:st_crs]{sf::st_crs()}} and \code{\link[terra:crs]{terra::crs()}}.

Coercing objects:
\code{\link[=as_coordinates]{as_coordinates()}},
\code{\link[=as_sf]{as_sf()}},
\code{\link[=as_spatraster]{as_spatraster()}},
\code{\link{as_tibble.Spat}},
\code{\link{fortify.Spat}},
\code{\link{tidy.Spat}}
}
\concept{coerce}
