% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-cols-SpatVector.R
\encoding{UTF-8}
\name{bind_cols.SpatVector}
\alias{bind_cols.SpatVector}
\alias{bind_spat_cols}
\title{Bind multiple \code{\link[terra:vect]{SpatVector}}, \code{\link[sf:st_sf]{sf}} and data frame
objects by column}
\usage{
bind_spat_cols(
  ...,
  .name_repair = c("unique", "universal", "check_unique", "minimal")
)
}
\arguments{
\item{...}{Objects to combine. The first argument must be a \code{SpatVector}.
Each subsequent argument can be a \code{SpatVector}, \code{\link[sf:st_sf]{sf}} object or
data frame.
Inputs are \link[vctrs:theory-faq-recycling]{recycled} to the same length,
then matched by position.}

\item{.name_repair}{One of \code{"unique"}, \code{"universal"}, or
\code{"check_unique"}. See \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}} for the meaning of these
options.}
}
\value{
A \code{\link[terra:vect]{SpatVector}} object with the combined columns. Its
geometry and
CRS come from the first \code{\link[terra:vect]{SpatVector}} in \code{...}.
}
\description{
Bind any number of \code{\link[terra:vect]{SpatVector}} objects, data frames and
\code{\link[sf:st_sf]{sf}} objects by
column,
making a wider result. This is similar to \code{do.call(cbind, data_frames)}.

Where possible, prefer using a \link[=mutate-joins.SpatVector]{join} to
combine \code{SpatVector} and data frame objects. \code{bind_spat_cols()}
binds rows in the order in which they appear, so it is easy to create
meaningless results without realizing it.
}
\section{\CRANpkg{terra} equivalent}{


\code{cbind()} method.
}

\section{Methods}{


Implementation of the \code{\link[dplyr:bind_cols]{dplyr::bind_cols()}} function for
\code{\link[terra:vect]{SpatVector}}
objects. For the second and subsequent arguments in \code{...}, the geometry is
not \code{cbind}ed and only the data frame-like columns are kept.
}

\examples{
library(terra)
sv <- vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))
df2 <- data.frame(letters = letters[seq_len(nrow(sv))])

# Data frame
bind_spat_cols(sv, df2)

# Another SpatVector
bind_spat_cols(sv[1:2, ], sv[3:4, ])

# sf objects
sfobj <- sf::read_sf(system.file("shape/nc.shp", package = "sf"))

bind_spat_cols(sv[1:9, ], sfobj[1:9, ])

# Mixed

end <- bind_spat_cols(sv, sfobj[seq_len(nrow(sv)), 1:2], df2)

end
glimpse(end)

# Row sizes must be compatible when column-binding.
try(bind_spat_cols(sv, sfobj))
}
\seealso{
\code{\link[dplyr:bind_cols]{dplyr::bind_cols()}}.

Other \CRANpkg{dplyr} verbs that operate on pairs of \code{SpatVector} and data frame objects:
\code{\link{bind_rows.SpatVector}},
\code{\link[=cross_join.SpatVector]{cross_join.SpatVector()}},
\code{\link{filter-joins.SpatVector}},
\code{\link{mutate-joins.SpatVector}},
\code{\link[=nest_join.SpatVector]{nest_join.SpatVector()}},
\code{\link{rows.SpatVector}}
}
\concept{dplyr.pairs}
