% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-rows-SpatVector.R
\encoding{UTF-8}
\name{bind_rows.SpatVector}
\alias{bind_rows.SpatVector}
\alias{bind_spat_rows}
\alias{bind.Spat}
\title{Bind multiple \code{\link[terra:vect]{SpatVector}}, \code{\link[sf:st_sf]{sf}}, \code{sfc} and data
frame objects by row}
\usage{
bind_spat_rows(..., .id = NULL)
}
\arguments{
\item{...}{Objects to combine. The first argument must be a \code{SpatVector}.
Each subsequent argument can be a \code{SpatVector}, \code{\link[sf:st_sf]{sf}} or \code{sfc}
object or a
data frame. Columns are matched by name and any missing columns are filled
with \code{NA}.}

\item{.id}{The name of an optional identifier column. Provide a string to
create an output column that identifies each input. The column will use
names if available, otherwise it will use positions.}
}
\value{
A \code{\link[terra:vect]{SpatVector}} object of the same type as the first
element of
\code{...}.
}
\description{
Bind any number of \code{\link[terra:vect]{SpatVector}} objects, data frames,
\code{\link[sf:st_sf]{sf}} and \code{sfc} objects
by row,
making a longer result. This is similar to \code{do.call(rbind, data_frames)},
but the output will contain all columns that appear in any of the inputs.
}
\section{\CRANpkg{terra} equivalent}{


\code{rbind()} method.
}

\section{Methods}{


Implementation of the \code{\link[dplyr:bind_rows]{dplyr::bind_rows()}} function for
\code{\link[terra:vect]{SpatVector}} objects.

The first argument should be a \code{\link[terra:vect]{SpatVector}}. Each subsequent
argument can be
a \code{SpatVector}, an \code{\link[sf:st_sf]{sf}} or \code{sfc} object or a data frame:
\itemize{
\item If subsequent spatial objects have a different CRS from the first element,
they are reprojected to the CRS of the first element with a message.
\item If any element of \code{...} is a \link[tibble:tbl_df]{tibble}/data frame, the rows
are column-bound
with empty geometries with a message.
}
}

\examples{

library(terra)
v <- vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))

v1 <- v[1, "cpro"]
v2 <- v[3:5, c("name", "iso2")]

# You can supply individual SpatVector objects as arguments.
bind_spat_rows(v1, v2)

# When you supply a column name with the `.id` argument, a new column is
# created to link each row to its original data frame.
bind_spat_rows(v1, v2, .id = "id")

\donttest{
# Use with sf objects.
sfobj <- sf::st_as_sf(v2[1, ])

sfobj

bind_spat_rows(v1, sfobj)

# Reproject with a message when the CRS differs.
sfobj_3857 <- as_spatvector(sfobj) |> project("EPSG:3857")

bind_spat_rows(v1, sfobj_3857)

# Bind data frames with a message.
data("mtcars")
bind_spat_rows(v1, sfobj, mtcars, .id = "id2")

# Use lists
bind_spat_rows(list(v1[1, ], sfobj[1:2, ]))

# Use a named list with `.id`.
bind_spat_rows(list(
  SpatVector = v1[1, ], sf = sfobj[1, ],
  mtcars = mtcars[1, ]
), .id = "source")
}
}
\seealso{
\code{\link[dplyr:bind_rows]{dplyr::bind_rows()}}.

Other \CRANpkg{dplyr} verbs that operate on pairs of \code{SpatVector} and data frame objects:
\code{\link{bind_cols.SpatVector}},
\code{\link[=cross_join.SpatVector]{cross_join.SpatVector()}},
\code{\link{filter-joins.SpatVector}},
\code{\link{mutate-joins.SpatVector}},
\code{\link[=nest_join.SpatVector]{nest_join.SpatVector()}},
\code{\link{rows.SpatVector}}
}
\concept{dplyr.pairs}
