% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coltab_palette.R
\name{coltab_palette}
\alias{coltab_palette}
\title{Extract a color palette from a color table associated with a SpatRaster}
\usage{
coltab_palette(x)
}
\arguments{
\item{x}{SpatRaster objects}
}
\value{
A named vector of hex colors or \code{NULL} if no color table is found
}
\description{
Some categorical SpatRaster may have an associated color table. This
function extract those values.
}
\section{terra equivalent}{


\code{\link[terra:colors]{terra::coltab()}}
}

\examples{
library(terra)
# Corine Land Cover v 2018
r <- rast(system.file("extdata/clc_2018_majorca.tif", package = "tidyterra"))

plot(r)
# Get coltab
coltab_pal <- coltab_palette(r)

scales::show_col(coltab_pal)
\donttest{
# With ggplot2 + tidyterra
library(ggplot2)

gg <- ggplot() +
  geom_spatraster(data = r) +
  guides(fill = guide_legend(ncol = 1, keyheight = 0.15)) +
  theme(legend.text = element_text(size = 3))


# Default plot
gg

# With coltabs
gg +
  scale_fill_manual(values = coltab_pal, na.value = NA)

}
}
\seealso{
\code{\link[terra:colors]{terra::coltab()}}

Other helpers: 
\code{\link{compare_spatrasters}()},
\code{\link{is_grouped_spatvector}()},
\code{\link{is_regular_grid}()},
\code{\link{pull_crs}()}
}
\concept{helpers}
