% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-spatrasters.R
\encoding{UTF-8}
\name{compare_spatrasters}
\alias{compare_spatrasters}
\title{Compare attributes of two \code{\link[terra:rast]{SpatRaster}} objects}
\usage{
compare_spatrasters(x, y, digits = 6)
}
\arguments{
\item{x, y}{\code{SpatRaster} objects.}

\item{digits}{Integer to set the precision for comparing the extent and
the resolution.}
}
\value{
An invisible logical value indicating whether the \code{\link[terra:rast]{SpatRaster}}
objects are compatible, plus an informative message flagging any issues
found.
}
\description{
Two \code{\link[terra:rast]{SpatRaster}} objects are compatible (in terms of combining
layers) if
the CRS, extent and resolution are similar. In those cases you can combine
the objects simply as \code{c(x, y)}.

This function compares those attributes and reports the results. See
\strong{Resolving differences} for minimal guidance.
}
\section{\CRANpkg{terra} equivalent}{


\code{\link[terra:compareGeom]{terra::compareGeom()}}.
}

\section{Resolving differences}{

\itemize{
\item For a \strong{different CRS}, try \code{\link[terra:project]{terra::project()}}.
\item For a \strong{different extent}, try \code{\link[terra:resample]{terra::resample()}}.
\item For a \strong{different resolution}, try \code{\link[terra:resample]{terra::resample()}},
\code{\link[terra:aggregate]{terra::aggregate()}} or \code{\link[terra:disagg]{terra::disagg()}}.
}
}

\examples{
library(terra)

x <- rast(matrix(1:90, ncol = 3), crs = "EPSG:3857")

# Nothing
compare_spatrasters(x, x)

# Different CRS
y_nocrs <- x
crs(y_nocrs) <- NA

compare_spatrasters(x, y_nocrs)

# Different extent
compare_spatrasters(x, x[1:10, , drop = FALSE])

# Different resolution
y_newres <- x

res(y_newres) <- res(x) / 2
compare_spatrasters(x, y_newres)

# Everything

compare_spatrasters(x, project(x, "epsg:3035"))

}
\seealso{
\code{\link[terra:identical]{terra::identical()}}.
}
\concept{helpers}
