% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete-SpatVector.R
\encoding{UTF-8}
\name{complete.SpatVector}
\alias{complete.SpatVector}
\title{Complete missing combinations in a \code{\link[terra:vect]{SpatVector}}}
\usage{
\method{complete}{SpatVector}(data, ..., fill = list(), explicit = TRUE)
}
\arguments{
\item{data}{A \code{SpatVector}.}

\item{...}{<\code{\link[tidyr:tidyr_data_masking]{data-masking}}> Specification of columns
to expand or complete. Columns can be atomic vectors or lists.
\itemize{
\item To find all unique combinations of \code{x}, \code{y} and \code{z}, including those not
present in the data, supply each variable as a separate argument:
\code{expand(df, x, y, z)} or \code{complete(df, x, y, z)}.
\item To find only the combinations that occur in the
data, use \code{nesting}: \code{expand(df, nesting(x, y, z))}.
\item You can combine the two forms. For example,
\code{expand(df, nesting(school_id, student_id), date)} would produce
a row for each present school-student combination for all possible
dates.
}

When used with factors, \code{\link[tidyr:expand]{expand()}} and \code{\link[tidyr:complete]{complete()}} use the full set of
levels, not just those that appear in the data. If you want to use only the
values seen in the data, use \code{forcats::fct_drop()}.

When used with continuous variables, you may need to fill in values
that do not appear in the data: to do so use expressions like
\code{year = 2010:2020} or \code{year = full_seq(year,1)}.}

\item{fill}{A named list that for each variable supplies a single value to
use instead of \code{NA} for missing combinations.}

\item{explicit}{Should both implicit (newly created) and explicit
(pre-existing) missing values be filled by \code{fill}? By default, this is
\code{TRUE}, but if set to \code{FALSE} this will limit the fill to only implicit
missing values.}
}
\value{
A \code{\link[terra:vect]{SpatVector}} object.
}
\description{
\code{complete()} turns implicit missing combinations in a
\code{\link[terra:vect]{SpatVector}}
into explicit rows while preserving geometry and spatial metadata.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[tidyr:complete]{tidyr::complete()}} method for
\code{\link[terra:vect]{SpatVector}} objects.

\code{complete()} preserves the geometry column while expanding missing
combinations. New combinations receive empty geometries.
}

\examples{
v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))
v <- dplyr::mutate(v, grp = ifelse(iso2 \%in\% c("ES-AV", "ES-BU"), "a", "b"))

complete(v, grp, tidyr::nesting(iso2, name)) |>
  glimpse()

}
\seealso{
\code{\link[tidyr:complete]{tidyr::complete()}}.

Other \CRANpkg{tidyr} verbs for handling missing values:
\code{\link{drop_na.Spat}},
\code{\link[=expand.SpatVector]{expand.SpatVector()}},
\code{\link[=fill.SpatVector]{fill.SpatVector()}},
\code{\link{replace_na.Spat}}
}
\concept{tidyr.missing}
