% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count-tally-SpatVector.R
\encoding{UTF-8}
\name{count.SpatVector}
\alias{count.SpatVector}
\alias{tally.SpatVector}
\alias{add_count.SpatVector}
\title{Count the observations in each \code{\link[terra:vect]{SpatVector}} group}
\usage{
\method{count}{SpatVector}(
  x,
  ...,
  wt = NULL,
  sort = FALSE,
  name = NULL,
  .drop = deprecated(),
  .dissolve = TRUE
)

\method{tally}{SpatVector}(x, wt = NULL, sort = FALSE, name = NULL)

\method{add_count}{SpatVector}(x, ..., wt = NULL, sort = FALSE, name = NULL, .drop = deprecated())
}
\arguments{
\item{x}{A \code{SpatVector} created with \code{\link[terra:vect]{terra::vect()}}.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Variables to group
by.}

\item{wt}{<\code{\link[rlang:args_data_masking]{data-masking}}> Frequency weights.
Can be \code{NULL} or a variable:
\itemize{
\item If \code{NULL} (the default), counts the number of rows in each group.
\item If a variable, computes \code{sum(wt)} for each group.
}}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}

\item{name}{The name of the new column in the output.

If omitted, it will default to \code{n}. If there's already a column called \code{n},
it will use \code{nn}. If there's a column called \code{n} and \code{nn}, it'll use
\code{nnn}, and so on, adding \code{n}s until it gets a new name.}

\item{.drop}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Argument no longer
supported, empty groups are always removed (see \code{\link[dplyr:count]{dplyr::count()}},
\code{.drop = TRUE} argument).}

\item{.dissolve}{Logical. If \code{TRUE}, dissolve borders between aggregated
geometries.}
}
\value{
A \code{\link[terra:vect]{SpatVector}} object with updated grouping metadata.
}
\description{
\code{count()} lets you quickly count the unique values of one or more variables:
\code{df |> count(a, b)} is roughly equivalent to
\code{df |> group_by(a, b) |> summarise(n = n())}. \code{count()} is paired with
\code{tally()}, a lower-level helper that is equivalent to
\code{df |> summarise(n = n())}. Supply \code{wt} to perform weighted counts,
switching the summary from \code{n = n()} to \code{n = sum(wt)}.

\code{add_count()} is equivalent to \code{count()} but uses \code{\link[=mutate]{mutate()}} instead of
\code{\link[=summarise]{summarise()}} so that it adds a new column with group-wise counts.
}
\section{\CRANpkg{terra} equivalent}{


\code{\link[terra:aggregate]{terra::aggregate()}}.
}

\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:count]{dplyr::count()}} method family for
\code{\link[terra:vect]{SpatVector}} objects.

\code{\link[=tally]{tally()}} will always return a disaggregated geometry while \code{\link[=count]{count()}} can
handle this. See also \code{\link[=summarise.SpatVector]{summarise.SpatVector()}}.
}

\examples{
\donttest{

library(terra)
f <- system.file("ex/lux.shp", package = "terra")
p <- vect(f)

p |> count(NAME_1, sort = TRUE)

p |> count(pop = ifelse(POP < 20000, "A", "B"))

# tally() is a lower-level function that assumes grouping is already done.
p |> tally()

p |>
  group_by(NAME_1) |>
  tally()

# Dissolve geometries by default

library(ggplot2)
p |>
  count(NAME_1) |>
  ggplot() +
  geom_spatvector(aes(fill = n))

# Opt out
p |>
  count(NAME_1, .dissolve = FALSE, sort = TRUE) |>
  ggplot() +
  geom_spatvector(aes(fill = n))
}
}
\seealso{
\code{\link[dplyr:count]{dplyr::count()}}, \code{\link[dplyr:tally]{dplyr::tally()}}.

Other \CRANpkg{dplyr} verbs that operate on groups of rows:
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}},
\code{\link[=reframe.SpatVector]{reframe.SpatVector()}},
\code{\link[=rowwise.SpatVector]{rowwise.SpatVector()}},
\code{\link[=summarise.SpatVector]{summarise.SpatVector()}}
}
\concept{dplyr.groups}
