% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-na-Spat.R
\encoding{UTF-8}
\name{drop_na.Spat}
\alias{drop_na.Spat}
\alias{drop_na.SpatVector}
\alias{drop_na.SpatRaster}
\title{Drop attributes of \verb{Spat*} objects containing missing values}
\usage{
\method{drop_na}{SpatVector}(data, ...)

\method{drop_na}{SpatRaster}(data, ...)
}
\arguments{
\item{data}{A \code{SpatVector} created with \code{\link[terra:vect]{terra::vect()}} or a \code{SpatRaster}
created with \code{\link[terra:rast]{terra::rast()}}.}

\item{...}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> Attributes to inspect
for missing values. If empty, all attributes are used.}
}
\value{
A \verb{Spat*} object of the same class as \code{data}. See \strong{Methods}.
}
\description{
\itemize{
\item \code{\link[terra:vect]{SpatVector}}: \code{drop_na()} drops geometries when any
attribute specified by \code{...} contains a missing value.
\item \code{\link[terra:rast]{SpatRaster}}: \code{drop_na()} drops cells when any layer
specified by \code{...} contains a missing value.
}
}
\section{\CRANpkg{terra} equivalent}{


\code{\link[terra:trim]{terra::trim()}}.
}

\section{Methods}{


Implementation of the \strong{generic} \code{\link[tidyr:drop_na]{tidyr::drop_na()}} methods for \verb{Spat*}
objects.
\subsection{\code{\link[terra:vect]{SpatVector}}}{

This method operates on attributes, meaning that \code{NA} values are assessed in
the attributes (columns) of each geometry (row). The result is a
\code{\link[terra:vect]{SpatVector}} with potentially fewer
geometries than the input.
}

\subsection{\code{\link[terra:rast]{SpatRaster}}}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}

The implementation of \code{\link[=drop_na.SpatRaster]{drop_na.SpatRaster()}} can be understood as a
masking method based on the values of the layers (see \code{\link[terra:mask]{terra::mask()}}).

\code{\link[terra:rast]{SpatRaster}} layers are treated as columns and \code{SpatRaster}
cells as
rows, so rows (cells) with any \code{NA} value on any layer become \code{NA}. You can
also mask the cells (rows) based on the values of specific layers (columns).

\code{drop_na()} effectively removes outer cells that are \code{NA} (see
\code{\link[terra:trim]{terra::trim()}}), so the extent of the resulting object may differ from the
extent of the input (see \code{\link[terra:resample]{terra::resample()}} for more information).

Check the \strong{Examples} to have a better understanding of this method.
\subsection{Feedback needed!}{

Visit \url{https://github.com/dieghernan/tidyterra/issues}. The implementation
of this method for \code{\link[terra:rast]{SpatRaster}} may change in the future.
}

}
}

\examples{

library(terra)

f <- system.file("extdata/cyl.gpkg", package = "tidyterra")

v <- terra::vect(f)

# Add missing values.
v <- v |> mutate(iso2 = ifelse(cpro <= "09", NA, cpro))

# Initial plot.
plot(v, col = "red")

# Drop geometries with missing values in iso2.
v |>
  drop_na(iso2) |>
  plot(col = "red")
# SpatRaster method

\donttest{
r <- rast(
  crs = "EPSG:3857",
  extent = c(0, 10, 0, 10),
  nlyr = 3,
  resolution = c(2.5, 2.5)
)
terra::values(r) <- seq_len(ncell(r) * nlyr(r))

# Add missing values.
r[r > 13 & r < 22 | r > 31 & r < 45] <- NA

# Initial plot.
plot(r, nc = 3)

# Mask with lyr.1.
r |>
  drop_na(lyr.1) |>
  plot(nc = 3)

# Mask with lyr.2.
r |>
  drop_na(lyr.2) |>
  plot(nc = 3)

# Mask with lyr.3.
r |>
  drop_na(lyr.3) |>
  plot(nc = 3)

# Mask all layers.
r |>
  drop_na() |>
  plot(nc = 3)
}
}
\seealso{
\code{\link[tidyr:drop_na]{tidyr::drop_na()}}.

Other \CRANpkg{tidyr} verbs for handling missing values:
\code{\link[=complete.SpatVector]{complete.SpatVector()}},
\code{\link[=expand.SpatVector]{expand.SpatVector()}},
\code{\link[=fill.SpatVector]{fill.SpatVector()}},
\code{\link{replace_na.Spat}}
}
\concept{tidyr.missing}
