% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-SpatVector.R
\encoding{UTF-8}
\name{expand.SpatVector}
\alias{expand.SpatVector}
\title{Expand \code{\link[terra:vect]{SpatVector}} attribute combinations}
\usage{
\method{expand}{SpatVector}(data, ..., .name_repair = "check_unique")
}
\arguments{
\item{data}{A \code{SpatVector}.}

\item{...}{<\code{\link[tidyr:tidyr_data_masking]{data-masking}}> Specification of columns
to expand or complete. Columns can be atomic vectors or lists.
\itemize{
\item To find all unique combinations of \code{x}, \code{y} and \code{z}, including those not
present in the data, supply each variable as a separate argument:
\code{expand(df, x, y, z)} or \code{complete(df, x, y, z)}.
\item To find only the combinations that occur in the
data, use \code{nesting}: \code{expand(df, nesting(x, y, z))}.
\item You can combine the two forms. For example,
\code{expand(df, nesting(school_id, student_id), date)} would produce
a row for each present school-student combination for all possible
dates.
}

When used with factors, \code{\link[tidyr:expand]{expand()}} and \code{\link[tidyr:complete]{complete()}} use the full set of
levels, not just those that appear in the data. If you want to use only the
values seen in the data, use \code{forcats::fct_drop()}.

When used with continuous variables, you may need to fill in values
that do not appear in the data: to do so use expressions like
\code{year = 2010:2020} or \code{year = full_seq(year,1)}.}

\item{.name_repair}{One of \code{"check_unique"}, \code{"unique"}, \code{"universal"},
\code{"minimal"}, \code{"unique_quiet"}, or \code{"universal_quiet"}. See \code{\link[vctrs:vec_as_names]{vec_as_names()}}
for the meaning of these options.}
}
\value{
A \link[tibble:tbl_df]{tibble}.
}
\description{
\code{expand()} returns a \link[tibble:tbl_df]{tibble} with all combinations of
selected attributes.
It does not return a \code{\link[terra:vect]{SpatVector}} because newly created
combinations do not
have a well-defined geometry. Use \code{\link[=complete.SpatVector]{complete.SpatVector()}} when empty
geometries should be added explicitly.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[tidyr:expand]{tidyr::expand()}} method for
\code{\link[terra:vect]{SpatVector}}
objects.

The output is a \link[tibble:tbl_df]{tibble} with attribute combinations.
Geometry is not
preserved because new combinations do not have a well-defined geometry.
}

\examples{
v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))
v$grp <- rep(c("A", "B"), length.out = nrow(v))

expand(v, grp, cpro)
}
\seealso{
\code{\link[tidyr:expand]{tidyr::expand()}}, \code{\link[=complete.SpatVector]{complete.SpatVector()}}.

Other \CRANpkg{tidyr} verbs for handling missing values:
\code{\link[=complete.SpatVector]{complete.SpatVector()}},
\code{\link{drop_na.Spat}},
\code{\link[=fill.SpatVector]{fill.SpatVector()}},
\code{\link{replace_na.Spat}}
}
\concept{tidyr.missing}
