% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-Spat.R
\encoding{UTF-8}
\name{filter.Spat}
\alias{filter.Spat}
\alias{filter.SpatRaster}
\alias{filter.SpatVector}
\alias{filter_out.SpatVector}
\title{Subset cells/geometries of \verb{Spat*} objects}
\usage{
\method{filter}{SpatRaster}(.data, ..., .preserve = FALSE, .keep_extent = TRUE)

\method{filter}{SpatVector}(.data, ..., .by = NULL, .preserve = FALSE)

\method{filter_out}{SpatVector}(.data, ..., .by = NULL, .preserve = FALSE)
}
\arguments{
\item{.data}{A \code{SpatRaster} created with \code{\link[terra:rast]{terra::rast()}} or a \code{SpatVector}
created with \code{\link[terra:vect]{terra::vect()}}.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Expressions that
return a logical value and are defined in terms of the layers/attributes
in \code{.data}. If multiple expressions are included, they are combined with
the \code{&} operator. Only cells/geometries for which all conditions evaluate
to \code{TRUE} are kept. See \strong{Methods}.}

\item{.preserve}{Relevant when the \code{.data} input is grouped. If \code{.preserve = FALSE} (the default), the grouping structure is recalculated based on the
resulting data, otherwise the grouping is kept as is.}

\item{.keep_extent}{Logical. If \code{TRUE}, keep the extent of the resulting
\code{\link[terra:rast]{SpatRaster}}. If \code{FALSE}, \code{\link[terra:trim]{terra::trim()}} is called so the
extent may
differ from the extent of the input. See also \code{\link[=drop_na.SpatRaster]{drop_na.SpatRaster()}}.}

\item{.by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, a selection of columns to
group by for just this operation, functioning as an alternative to \code{\link[dplyr:group_by]{group_by()}}. For
details and examples, see \link[dplyr:dplyr_by]{?dplyr_by}.}
}
\value{
A \verb{Spat*} object of the same class as \code{.data}. See \strong{Methods}.
}
\description{
These functions subset a \verb{Spat*} object by applying the expressions in \code{...}
to determine which rows should be kept (for \code{filter()}) or dropped (for
\code{filter_out()}).

Multiple conditions can be supplied separated by a comma. These will be
combined with the \code{&} operator. To combine comma-separated conditions using
\code{|} instead, wrap them in \code{\link[dplyr:when_any]{dplyr::when_any()}}.

Both \code{filter()} and \code{filter_out()} treat \code{NA} like \code{FALSE}. This subtle
behavior can affect how you write your conditions when missing values are
involved. See \code{\link[dplyr:filter]{dplyr::filter()}}.

\strong{You can filter a \code{\link[terra:rast]{SpatRaster}} by its geographic
coordinates}.
Use \code{filter(.data, x > 42)}. The names \code{x} and \code{y} are reserved in
\CRANpkg{terra} because they refer to the geographic coordinates of the
layer.

See \strong{Examples} and section \strong{Layer and column names} on
\code{\link[=as_tibble.Spat]{as_tibble.Spat()}}.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:filter]{dplyr::filter()}} methods for \verb{Spat*}
objects.
\subsection{\code{\link[terra:rast]{SpatRaster}}}{

Cells that do not meet the conditions in \code{...} are returned as \code{NA}. On a
multi-layer \code{\link[terra:rast]{SpatRaster}}, \code{NA} is propagated across all
layers.

If \code{.keep_extent = TRUE}, the returned \code{\link[terra:rast]{SpatRaster}} has the
same CRS,
extent, resolution and number of cells as \code{.data}. If
\code{.keep_extent = FALSE}, the outer \code{NA} cells are trimmed with
\code{\link[terra:trim]{terra::trim()}}, so the extent and number of cells may differ. The output
still has the same CRS and resolution as \code{.data}.

The \code{x} and \code{y} coordinates of the \code{\link[terra:rast]{SpatRaster}} are also
available
internally for filtering. See \strong{Examples}.
}

\subsection{\code{\link[terra:vect]{SpatVector}}}{

The result is a \code{\link[terra:vect]{SpatVector}} containing the geometries whose
attributes
satiy all conditions.
}
}

\examples{

library(terra)
f <- system.file("extdata/cyl_temp.tif", package = "tidyterra")

r <- rast(f) |> select(tavg_04)

plot(r)

# Filter temps
r_f <- r |> filter(tavg_04 > 11.5)

# Extent is kept
plot(r_f)

# Filter temps and extent
r_f2 <- r |> filter(tavg_04 > 11.5, .keep_extent = FALSE)

# Extent has changed
plot(r_f2)

# Filter by geographic coordinates
r2 <- project(r, "epsg:4326")

r2 |> plot()

r2 |>
  filter(
    x > -4,
    x < -2,
    y > 42
  ) |>
  plot()
v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))
glimpse(v)
v |> filter(cpro < 10)

# Same as
v |> filter_out(cpro >= 10)
}
\seealso{
\code{\link[dplyr:filter]{dplyr::filter()}}.

Other \CRANpkg{dplyr} verbs that operate on rows:
\code{\link[=arrange.SpatVector]{arrange.SpatVector()}},
\code{\link[=distinct.SpatVector]{distinct.SpatVector()}},
\code{\link{rows.SpatVector}},
\code{\link{slice.Spat}}
}
\concept{dplyr.rows}
