% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-Spat.R
\encoding{UTF-8}
\name{fortify.Spat}
\alias{fortify.Spat}
\alias{fortify.SpatRaster}
\alias{fortify.SpatVector}
\alias{fortify.SpatGraticule}
\alias{fortify.SpatExtent}
\title{Fortify \verb{Spat*} objects}
\usage{
\method{fortify}{SpatRaster}(
  model,
  data,
  ...,
  .name_repair = c("unique", "check_unique", "universal", "minimal", "unique_quiet",
    "universal_quiet"),
  maxcell = terra::ncell(model) * 1.1,
  pivot = FALSE
)

\method{fortify}{SpatVector}(model, data, ...)

\method{fortify}{SpatGraticule}(model, data, ...)

\method{fortify}{SpatExtent}(model, data, ..., crs = "")
}
\arguments{
\item{model}{A \code{SpatRaster} created with \code{\link[terra:rast]{terra::rast()}}, a \code{SpatVector}
created with \code{\link[terra:vect]{terra::vect()}}, a \code{SpatGraticule} (see \code{\link[terra:graticule]{terra::graticule()}})
or a \code{SpatExtent} (see \code{\link[terra:ext]{terra::ext()}}).}

\item{data}{Not used by this method.}

\item{...}{Ignored by these methods.}

\item{.name_repair}{Treatment of problematic column names:
\itemize{
\item \code{"minimal"}: No name repair or checks, beyond basic existence,
\item \code{"unique"}: Make sure names are unique and not empty,
\item \code{"check_unique"}: (default value), no name repair, but check they are
\code{unique},
\item \code{"universal"}: Make the names \code{unique} and syntactic
\item \code{"unique_quiet"}: Same as \code{"unique"}, but "quiet"
\item \code{"universal_quiet"}: Same as \code{"universal"}, but "quiet"
\item a function: apply custom name repair (e.g., \code{.name_repair = make.names}
for names in the style of base R).
\item A purrr-style anonymous function, see \code{\link[rlang:as_function]{rlang::as_function()}}
}

This argument is passed on as \code{repair} to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
See there for more details on these terms and the strategies used
to enforce them.}

\item{maxcell}{Positive integer or \code{Inf}. Maximum number of raster cells to
use. Use \code{Inf} to use all cells.}

\item{pivot}{Logical. When \code{TRUE}, a \code{SpatRaster} is returned in \link[tidyr:pivot_longer]{long format}. When \code{FALSE} (the default), it is returned
as a data frame with one column per layer. See \strong{Details}.}

\item{crs}{Input that includes or represents a CRS. It can be an \code{sf} or
\code{sfc} object, a \code{\link[terra:rast]{SpatRaster}} or \code{\link[terra:vect]{SpatVector}}
object, a \code{crs} object from
\code{\link[sf:st_crs]{sf::st_crs()}}, a character string (for example a \href{https://proj.org/en/9.3/operations/projections/index.html}{PROJ string}) or
an integer representing an \href{https://epsg.io/}{EPSG} code.}
}
\value{
\code{\link[=fortify.SpatVector]{fortify.SpatVector()}}, \code{\link[=fortify.SpatGraticule]{fortify.SpatGraticule()}} and \code{\link[=fortify.SpatExtent]{fortify.SpatExtent()}}
return an \code{\link[sf:st_sf]{sf}} object.

\code{\link[=fortify.SpatRaster]{fortify.SpatRaster()}} returns a \link[tibble:tbl_df]{tibble}. See \strong{Methods}.
}
\description{
Fortify \code{\link[terra:rast]{SpatRaster}} and \code{\link[terra:vect]{SpatVector}} objects
to data frames. This provides
native compatibility with \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.

These methods are now implemented as wrappers around \code{\link{tidy.Spat}} methods.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[ggplot2:fortify]{ggplot2::fortify()}} methods for \verb{Spat*}
objects.
\subsection{\code{\link[terra:rast]{SpatRaster}}}{

Returns a \link[tibble:tbl_df]{tibble} that can be used with \verb{ggplot2::geom_*},
such as
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} and \code{\link[ggplot2:geom_raster]{ggplot2::geom_raster()}}.

The resulting \link[tibble:tbl_df]{tibble} includes coordinates in the \code{x} and
\code{y} columns. The
values of each layer are added as extra columns using the layer names from
the \code{\link[terra:rast]{SpatRaster}}.

The CRS of the \code{\link[terra:rast]{SpatRaster}} can be retrieved with
\code{attr(fortifiedSpatRaster, "crs")}.

You can convert the fortified object back to a \code{\link[terra:rast]{SpatRaster}}
with
\code{\link[=as_spatraster]{as_spatraster()}}.

When \code{pivot = TRUE}, the \code{\link[terra:rast]{SpatRaster}} is fortified in long
format (see
\code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}). The fortified object has the following columns:
\itemize{
\item \code{x}, \code{y}: Coordinates of the cell center in the corresponding CRS.
\item \code{lyr}: Name of the \code{\link[terra:rast]{SpatRaster}} layer associated with
\code{value}.
\item \code{value}: Cell value for the corresponding \code{lyr}.
}

This option can be useful when combining several \verb{geom_*} layers or when
faceting.
}

\subsection{\code{\link[terra:vect]{SpatVector}}, \code{SpatGraticule} and \code{SpatExtent}}{

Returns an \code{\link[sf:st_sf]{sf}} object that can be used with
\code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}}.
For \code{SpatVector} objects, Z and M coordinates are dropped for plotting.
Use \code{\link[=as_sf]{as_sf()}} to preserve these coordinates when converting to \code{sf}.
}
}

\examples{
\donttest{

# Demonstrate use with ggplot2.
library(ggplot2)

# Get a SpatRaster.
r <- system.file("extdata/volcano2.tif", package = "tidyterra") |>
  terra::rast() |>
  terra::project("EPSG:4326")

# You can now use a SpatRaster with any geom.
ggplot(r, maxcell = 50) +
  geom_histogram(aes(x = elevation),
    bins = 20, fill = "lightblue",
    color = "black"
  )

# For SpatVector, SpatGraticule and SpatExtent, use geom_sf().

# Create a SpatVector.
extfile <- system.file("extdata/cyl.gpkg", package = "tidyterra")
cyl <- terra::vect(extfile)

class(cyl)

ggplot(cyl) +
  geom_sf()

# SpatGraticule
g <- terra::graticule(60, 30, crs = "+proj=robin")

class(g)

ggplot(g) +
  geom_sf()

# SpatExtent
ex <- terra::ext(cyl)

class(ex)

ggplot(ex, crs = cyl) +
  geom_sf(fill = "red", alpha = 0.3) +
  geom_sf(data = cyl, fill = NA)
}

}
\seealso{
\code{\link{tidy.Spat}}, \code{\link[sf:st_as_sf]{sf::st_as_sf()}}.

Other \CRANpkg{ggplot2} helpers:
\code{\link{autoplot.Spat}},
\code{\link{geom_spat_contour}},
\code{\link[=geom_spatraster]{geom_spatraster()}},
\code{\link[=geom_spatraster_rgb]{geom_spatraster_rgb()}},
\code{\link{ggspatvector}},
\code{\link[=stat_spat_coordinates]{stat_spat_coordinates()}}

Other \CRANpkg{ggplot2} methods:
\code{\link{autoplot.Spat}}

Coercing objects:
\code{\link[=as_coordinates]{as_coordinates()}},
\code{\link[=as_sf]{as_sf()}},
\code{\link[=as_spatraster]{as_spatraster()}},
\code{\link[=as_spatvector]{as_spatvector()}},
\code{\link{as_tibble.Spat}},
\code{\link{tidy.Spat}}
}
\concept{coerce}
\concept{ggplot2.methods}
\concept{ggplot2.utils}
