% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-spatraster-rgb.R
\encoding{UTF-8}
\name{geom_spatraster_rgb}
\alias{geom_spatraster_rgb}
\title{Plot \code{\link[terra:rast]{SpatRaster}} objects as images}
\source{
Based on the \code{layer_spatial()} implementation in the \CRANpkg{ggspatial}
package. Thanks to \href{https://github.com/paleolimbot}{Dewey Dunnington} and to
\href{https://github.com/paleolimbot/ggspatial/graphs/contributors}{ggspatial contributors}.
}
\usage{
geom_spatraster_rgb(
  mapping = aes(),
  data,
  interpolate = TRUE,
  r = 1,
  g = 2,
  b = 3,
  alpha = 1,
  maxcell = 5e+05,
  max_col_value = 255,
  ...,
  stretch = NULL,
  zlim = NULL,
  mask_projection = FALSE
)
}
\arguments{
\item{mapping}{Ignored.}

\item{data}{A \code{SpatRaster} object.}

\item{interpolate}{If \code{TRUE} interpolate linearly, if \code{FALSE}
(the default) don't interpolate.}

\item{r, g, b}{Integer giving the layer number in \code{data} used for the red
(\code{r}), green (\code{g}) and blue (\code{b}) channel.}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{maxcell}{Positive integer or \code{Inf}. Maximum number of raster cells to
use. Use \code{Inf} to use all cells.}

\item{max_col_value}{Number giving the upper bound of the color value range.
When this is \code{255} (the default), the result is computed most efficiently.
See \code{\link[grDevices:rgb]{grDevices::rgb()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{stretch}{character. Option to stretch the values to increase contrast: "lin" (linear) or "hist" (histogram). The linear stretch uses \code{\link[terra]{stretch}} with arguments \code{minq=0.02} and \code{maxq=0.98}}

\item{zlim}{numeric vector of length 2. Range of values to plot (optional). If this is set, and \code{stretch="lin"} is used, then the values are stretched within the range of \code{zlim}. This allows creating consistent coloring between SpatRasters with different cell-value ranges, even when stretching the colors for improved contrast}

\item{mask_projection}{Logical, defaults to \code{FALSE}. If \code{TRUE}, mask out
areas outside the input extent. For example, to avoid data wrapping
around the dateline in equal-area projections. This argument is passed
to \code{\link[terra:project]{terra::project()}} when reprojecting the \code{\link[terra:rast]{SpatRaster}}.}
}
\value{
A \CRANpkg{ggplot2} layer.
}
\description{
This geom plots \code{\link[terra:rast]{SpatRaster}} objects (see \code{\link[terra:rast]{terra::rast()}}) as
RGB images.
The layers are combined so they represent the red, green and blue channels.

For plotting \code{\link[terra:rast]{SpatRaster}} objects by layer values use
\code{\link[=geom_spatraster]{geom_spatraster()}}.

The underlying implementation is based on \code{\link[ggplot2:geom_raster]{ggplot2::geom_raster()}}.
}
\section{\CRANpkg{terra} equivalent}{


\code{\link[terra:plotRGB]{terra::plotRGB()}}.
}

\section{Aesthetics}{


No \code{aes()} is required. In fact, \code{aes()} will be ignored.
}

\section{Coordinates}{


When the \code{\link[terra:rast]{SpatRaster}} does not have a CRS, that is,
\code{terra::crs(rast) == ""}, the geom does not make any assumption about the
scales.

On \code{\link[terra:rast]{SpatRaster}} objects that have a CRS, the geom uses
\code{\link[ggplot2:coord_sf]{ggplot2::coord_sf()}} to adjust the scales. This means that the
\strong{\code{SpatRaster} may be reprojected}.
}

\examples{
\donttest{

# Tile of Castile and Leon (Spain) from OpenStreetMap
file_path <- system.file("extdata/cyl_tile.tif", package = "tidyterra")

library(terra)
tile <- rast(file_path)

library(ggplot2)

ggplot() +
  geom_spatraster_rgb(data = tile) +
  # You can use coord_sf
  coord_sf(crs = 3035)

# Combine with sf objects
vect_path <- system.file("extdata/cyl.gpkg", package = "tidyterra")

cyl_sf <- sf::st_read(vect_path)

ggplot(cyl_sf) +
  geom_spatraster_rgb(data = tile) +
  geom_sf(aes(fill = iso2)) +
  coord_sf(crs = 3857) +
  scale_fill_viridis_d(alpha = 0.7)
}
}
\seealso{
\code{\link[ggplot2:geom_raster]{ggplot2::geom_raster()}}, \code{\link[ggplot2:coord_sf]{ggplot2::coord_sf()}}, \code{\link[grDevices:rgb]{grDevices::rgb()}}.

You can also get RGB tiles from the \CRANpkg{maptiles} package. See
\code{\link[maptiles:get_tiles]{maptiles::get_tiles()}}.

Other \CRANpkg{ggplot2} helpers:
\code{\link{autoplot.Spat}},
\code{\link{fortify.Spat}},
\code{\link{geom_spat_contour}},
\code{\link[=geom_spatraster]{geom_spatraster()}},
\code{\link{ggspatvector}},
\code{\link[=stat_spat_coordinates]{stat_spat_coordinates()}}
}
\concept{ggplot2.utils}
