% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-spatvector.R
\encoding{UTF-8}
\name{ggspatvector}
\alias{ggspatvector}
\alias{geom_spatvector}
\alias{geom_spatvector_label}
\alias{geom_spatvector_text}
\alias{stat_spatvector}
\title{Plot \code{\link[terra:vect]{SpatVector}} objects}
\usage{
geom_spatvector(
  mapping = aes(),
  data = NULL,
  na.rm = FALSE,
  show.legend = NA,
  ...
)

geom_spatvector_label(
  mapping = aes(),
  data = NULL,
  na.rm = FALSE,
  show.legend = NA,
  ...,
  inherit.aes = TRUE
)

geom_spatvector_text(
  mapping = aes(),
  data = NULL,
  na.rm = FALSE,
  show.legend = NA,
  ...,
  check_overlap = FALSE,
  inherit.aes = TRUE
)

stat_spatvector(
  mapping = NULL,
  data = NULL,
  geom = "rect",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{A \code{SpatVector} object, see \code{\link[terra:vect]{terra::vect()}}.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.

You can also set this to one of "polygon", "line", and "point" to
override the default legend.}

\item{...}{Other arguments passed to \code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}} functions.
These are often aesthetics, used to set an aesthetic to a fixed value,
like \code{colour = "red"} or \code{linewidth = 3}.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{check_overlap}{If \code{TRUE}, text that overlaps previous text in the
same layer will not be plotted. \code{check_overlap} happens at draw time and in
the order of the data. Therefore data should be arranged by the label
column before calling \code{geom_text()}. Note that this argument is not
supported by \code{geom_label()}.}

\item{geom}{The geometric object to use to display the data for this layer.
When using a \verb{stat_*()} function to construct a layer, the \code{geom} argument
can be used to override the default coupling between stats and geoms. The
\code{geom} argument accepts the following:
\itemize{
\item A \code{Geom} ggproto subclass, for example \code{GeomPoint}.
\item A string naming the geom. To give the geom as a string, strip the
function name of the \code{geom_} prefix. For example, to use \code{geom_point()},
give the geom as \code{"point"}.
\item For more information and other ways to specify the geom, see the
\link[ggplot2:layer_geoms]{layer geom} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}
}
\value{
A \CRANpkg{ggplot2} layer.
}
\description{
Wrappers of the \code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}} family used to plot
\code{\link[terra:vect]{SpatVector}}
objects (see \code{\link[terra:vect]{terra::vect()}}).
}
\details{
These functions are wrappers of \code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}} functions. Since a
\code{\link[=fortify.SpatVector]{fortify.SpatVector()}} method is provided, \CRANpkg{ggplot2} treats a
\code{\link[terra:vect]{SpatVector}} in the same way as an \code{\link[sf:st_sf]{sf}} object.
A side effect is that you can use \code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}} directly with
\code{\link[terra:vect]{SpatVector}} objects.

See \code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}} for details on aesthetics, etc.
}
\section{\CRANpkg{terra} equivalent}{


\code{\link[terra:plot]{terra::plot()}}.
}

\examples{
\donttest{
# Create a SpatVector
extfile <- system.file("extdata/cyl.gpkg", package = "tidyterra")

cyl <- terra::vect(extfile)
class(cyl)

library(ggplot2)

ggplot(cyl) +
  geom_spatvector()

# With params

ggplot(cyl) +
  geom_spatvector(aes(fill = name), color = NA) +
  scale_fill_viridis_d() +
  coord_sf(crs = 3857)

# Add labels
ggplot(cyl) +
  geom_spatvector(aes(fill = name), color = NA) +
  geom_spatvector_text(aes(label = iso2),
    fontface = "bold",
    color = "red"
  ) +
  scale_fill_viridis_d(alpha = 0.4) +
  coord_sf(crs = 3857)

# You can now use geom_sf() with SpatVector objects.

ggplot(cyl) +
  geom_sf() +
  labs(
    title = paste("cyl is", as.character(class(cyl))),
    subtitle = "With geom_sf()"
  )
}
}
\seealso{
\code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}}.

Other \CRANpkg{ggplot2} helpers:
\code{\link{autoplot.Spat}},
\code{\link{fortify.Spat}},
\code{\link{geom_spat_contour}},
\code{\link[=geom_spatraster]{geom_spatraster()}},
\code{\link[=geom_spatraster_rgb]{geom_spatraster_rgb()}},
\code{\link[=stat_spat_coordinates]{stat_spat_coordinates()}}
}
\concept{ggplot2.utils}
