% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grass_db}
\alias{grass_db}
\title{GRASS color tables}
\format{
A \link[tibble:tbl_df]{tibble} of 2920 rows and 6
columns with
the following fields:

\describe{
\item{pal}{Name of the palette.}
\item{limit}{(Optional) limit for each color.}
\item{r}{Value of the red channel (RGB color mode).}
\item{g}{Value of the green channel (RGB color mode).}
\item{b}{Value of the blue channel (RGB color mode).}
\item{hex}{Hex code of the color.}
}
}
\source{
Derived from \url{https://github.com/OSGeo/grass/tree/main/lib/gis/colors}. See
also \href{https://grass.osgeo.org/grass-stable/manuals/r.colors.html}{r.color - GRASS GIS Manual}.
}
\description{
A \link[tibble:tbl_df]{tibble} including the color map of
51 gradient palettes. Some palettes also
include a definition of color limits that can be used with
\code{\link[ggplot2:scale_fill_gradientn]{ggplot2::scale_fill_gradientn()}}.
}
\details{
Summary of palettes provided, description and recommended use:\tabular{llll}{
   \strong{palette} \tab \strong{use} \tab \strong{description} \tab \strong{range} \cr
   \code{aspect} \tab General \tab Aspect-oriented gray colors \tab  \cr
   \code{aspectcolr} \tab General \tab Aspect-oriented rainbow colors \tab 0 to 360 \cr
   \code{bcyr} \tab General \tab Blue through cyan and yellow to red \tab  \cr
   \code{bgyr} \tab General \tab Blue through green and yellow to red \tab  \cr
   \code{blues} \tab General \tab White to blue \tab  \cr
   \code{byg} \tab General \tab Blue through yellow to green \tab  \cr
   \code{byr} \tab General \tab Blue through yellow to red \tab  \cr
   \code{celsius} \tab General \tab Blue to red for Celsius temperatures \tab -80 to 80 \cr
   \code{corine} \tab Land cover \tab EU Corine land cover colors \tab 111 to 995 \cr
   \code{curvature} \tab General \tab Terrain curvature colors \tab -0.1 to 0.1 \cr
   \code{differences} \tab General \tab Difference-oriented colors \tab  \cr
   \code{elevation} \tab Topography \tab Relative raster values mapped to elevation colors \tab  \cr
   \code{etopo2} \tab Topography \tab ETOPO2 worldwide bathymetry and topography colors \tab -11000 to 8850 \cr
   \code{evi} \tab Natural \tab Enhanced Vegetation Index colors \tab -1 to 1 \cr
   \code{fahrenheit} \tab Temperature \tab Blue to red for Fahrenheit temperatures \tab -112 to 176 \cr
   \code{forest_cover} \tab Natural \tab Percentage of forest cover \tab 0 to 1 \cr
   \code{gdd} \tab Natural \tab Accumulated growing degree days \tab 0 to 6000 \cr
   \code{grass} \tab General \tab Perceptually uniform GRASS GIS green \tab  \cr
   \code{greens} \tab General \tab White to green \tab  \cr
   \code{grey} \tab General \tab Grayscale \tab  \cr
   \code{gyr} \tab General \tab Green through yellow to red \tab  \cr
   \code{haxby} \tab Topography \tab Relative colors for bathymetry or topography \tab  \cr
   \code{inferno} \tab General \tab Inferno perceptually uniform sequential color table \tab  \cr
   \code{kelvin} \tab Temperature \tab Blue to red for temperatures in Kelvin \tab 193.15 to 353.15 \cr
   \code{magma} \tab General \tab Magma perceptually uniform sequential color table \tab  \cr
   \code{ndvi} \tab Natural \tab Normalized Difference Vegetation Index colors \tab -1 to 1 \cr
   \code{ndwi} \tab Natural \tab Normalized Difference Water Index colors \tab -200 to 200 \cr
   \code{nlcd} \tab Land cover \tab US National Land Cover Dataset colors \tab 0 to 95 \cr
   \code{oranges} \tab General \tab White to orange \tab  \cr
   \code{plasma} \tab General \tab Plasma perceptually uniform sequential color table \tab  \cr
   \code{population} \tab Human \tab Human population classification breaks \tab 0 to 1000000 \cr
   \code{population_dens} \tab Human \tab Human population density classification breaks \tab 0 to 1000 \cr
   \code{precipitation} \tab Climate \tab Precipitation color table, 0 to 2000 mm \tab 0 to 7000 \cr
   \code{precipitation_daily} \tab Climate \tab Daily precipitation color table, 0 to 1000 mm \tab 0 to 100 \cr
   \code{precipitation_monthly} \tab Climate \tab Monthly precipitation color table, 0 to 1000 mm \tab 0 to 1000 \cr
   \code{rainbow} \tab General \tab Rainbow color table \tab  \cr
   \code{ramp} \tab General \tab Color ramp \tab  \cr
   \code{reds} \tab General \tab White to red \tab  \cr
   \code{roygbiv} \tab General \tab  \tab  \cr
   \code{rstcurv} \tab General \tab Terrain curvature from r.resamp.rst \tab -0.1 to 0.1 \cr
   \code{ryb} \tab General \tab Red through yellow to blue \tab  \cr
   \code{ryg} \tab General \tab Red through yellow to green \tab  \cr
   \code{sepia} \tab General \tab Yellowish-brown to white \tab  \cr
   \code{slope} \tab General \tab \code{r.slope.aspect}-style slope colors for raster values from 0 to 90 \tab 0 to 90 \cr
   \code{soilmoisture} \tab Natural \tab Soil moisture color table, 0.0 to 1.0 \tab 0 to 1 \cr
   \code{srtm} \tab Topography \tab Shuttle Radar Topography Mission elevation colors \tab -11000 to 8850 \cr
   \code{srtm_plus} \tab Topography \tab Shuttle Radar Topography Mission elevation colors with seafloor colors \tab -11000 to 8850 \cr
   \code{terrain} \tab Topography \tab Global elevation color table from -11000 to +8850 m \tab -11000 to 8850 \cr
   \code{viridis} \tab General \tab Viridis perceptually uniform sequential color table \tab  \cr
   \code{water} \tab Natural \tab Water depth \tab  \cr
   \code{wave} \tab General \tab Color wave \tab  \cr
}
}
\section{\CRANpkg{terra} equivalent}{


\code{\link[terra:map.pal]{terra::map.pal()}}.
}

\examples{
\donttest{
data("grass_db")

grass_db
# Select a palette

srtm_plus <- grass_db |>
  filter(pal == "srtm_plus")

f <- system.file("extdata/asia.tif", package = "tidyterra")
r <- terra::rast(f)

library(ggplot2)

p <- ggplot() +
  geom_spatraster(data = r) +
  labs(fill = "elevation")

p +
  scale_fill_gradientn(colors = srtm_plus$hex)

# Use with limits
p +
  scale_fill_gradientn(
    colors = srtm_plus$hex,
    values = scales::rescale(srtm_plus$limit),
    limit = range(srtm_plus$limit),
    na.value = "lightblue"
  )
}
}
\references{
GRASS Development Team (2024). \emph{Geographic Resources Analysis Support System
(GRASS) Software, Version 8.3.2}. Open Source Geospatial Foundation, USA.
\url{https://grass.osgeo.org}.
}
\seealso{
\code{\link[=scale_fill_grass_c]{scale_fill_grass_c()}}.

Other datasets:
\code{\link{cross_blended_hypsometric_tints_db}},
\code{\link{hypsometric_tints_db}},
\code{\link{princess_db}},
\code{\link{volcano2}}
}
\concept{datasets}
