% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by-SpatVector.R
\name{group-by.SpatVector}
\alias{group-by.SpatVector}
\alias{group_by.SpatVector}
\title{TODO}
\usage{
\method{group_by}{SpatVector}(.data, ..., .add = FALSE, .drop = group_by_drop_default(.data))
}
\description{
TODO
}
\seealso{
\code{\link[dplyr:group_by]{dplyr::group_by()}}, \code{\link[dplyr:group_by]{dplyr::ungroup()}}

Other dplyr methods:
\code{\link{arrange.SpatVector}()},
\code{\link{distinct.SpatVector}()},
\code{\link{filter-joins.SpatVector}},
\code{\link{filter.Spat}},
\code{\link{glimpse.Spat}},
\code{\link{mutate-joins.SpatVector}},
\code{\link{mutate.Spat}},
\code{\link{pull.Spat}},
\code{\link{relocate.Spat}},
\code{\link{rename.Spat}},
\code{\link{select.Spat}},
\code{\link{slice.Spat}}

Other single table verbs: 
\code{\link{arrange.SpatVector}()},
\code{\link{distinct.SpatVector}()},
\code{\link{filter.Spat}},
\code{\link{glimpse.Spat}},
\code{\link{mutate.Spat}},
\code{\link{rename.Spat}},
\code{\link{select.Spat}},
\code{\link{slice.Spat}}
}
\concept{dplyr.methods}
\concept{single table verbs}
