% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-trim.SpatVector.R
\encoding{UTF-8}
\name{group-trim.SpatVector}
\alias{group-trim.SpatVector}
\alias{group_trim.SpatVector}
\title{Trim grouping structure}
\usage{
\method{group_trim}{SpatVector}(.tbl, .drop = group_by_drop_default(.tbl))
}
\arguments{
\item{.tbl}{A \code{SpatVector} object. See \strong{Methods}.}

\item{.drop}{See \code{\link[=group_by.SpatVector]{group_by.SpatVector()}}.}
}
\value{
A \code{SpatVector} object with an additional attribute.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Drop unused levels of all factors that are used as grouping variables,
then recalculates the grouping structure.

\code{group_trim()} is particularly useful after a \code{\link[=filter]{filter()}} that is intended
to select a subset of groups.
}
\details{
See \strong{Details} on \code{\link[dplyr:group_by]{dplyr::group_by()}}.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:group_trim]{dplyr::group_trim()}} for \code{SpatVector}
objects.
}

\examples{
v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))

v$group <- rep(c("A", "B", "C"), 3)

v |>
  group_by(group) |>
  filter(group == "B", .preserve = TRUE) |>
  group_trim()

}
\seealso{
\code{\link[dplyr:group_trim]{dplyr::group_trim()}}.

Other \CRANpkg{dplyr} verbs that operate on group of rows:
\code{\link[=count.SpatVector]{count.SpatVector()}},
\code{\link{group-by.SpatVector}},
\code{\link[=rowwise.SpatVector]{rowwise.SpatVector()}},
\code{\link[=summarise.SpatVector]{summarise.SpatVector()}}

Other \CRANpkg{dplyr} methods:
\code{\link[=arrange.SpatVector]{arrange.SpatVector()}},
\code{\link{bind_cols.SpatVector}},
\code{\link{bind_rows.SpatVector}},
\code{\link[=count.SpatVector]{count.SpatVector()}},
\code{\link{cross-join.SpatVector}},
\code{\link[=distinct.SpatVector]{distinct.SpatVector()}},
\code{\link{filter-joins.SpatVector}},
\code{\link{filter.Spat}},
\code{\link{glimpse.Spat}},
\code{\link{group-by.SpatVector}},
\code{\link{mutate-joins.SpatVector}},
\code{\link{mutate.Spat}},
\code{\link{pull.Spat}},
\code{\link{relocate.Spat}},
\code{\link{rename.Spat}},
\code{\link[=rowwise.SpatVector]{rowwise.SpatVector()}},
\code{\link{select.Spat}},
\code{\link{slice.Spat}},
\code{\link[=summarise.SpatVector]{summarise.SpatVector()}}
}
\concept{dplyr.groups}
\concept{dplyr.methods}
