% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by-SpatVector.R
\encoding{UTF-8}
\name{group_by.SpatVector}
\alias{group_by.SpatVector}
\alias{ungroup.SpatVector}
\title{Group a \code{\link[terra:vect]{SpatVector}} by one or more variables}
\usage{
\method{group_by}{SpatVector}(.data, ..., .add = FALSE, .drop = group_by_drop_default(.data))

\method{ungroup}{SpatVector}(x, ...)
}
\arguments{
\item{.data, x}{A \code{SpatVector} object. See \strong{Methods}.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> In \code{group_by()},
variables or computations to group by. Computations are always done on the
ungrouped data frame. To perform computations on the grouped data, you need
to use a separate \code{mutate()} step before the \code{group_by()}.
Computations are not allowed in \code{nest_by()}.
In \code{ungroup()}, variables to remove from the grouping.}

\item{.add}{When \code{FALSE}, the default, \code{group_by()} will
override existing groups. To add to the existing groups, use
\code{.add = TRUE}.}

\item{.drop}{Drop groups formed by factor levels that don't appear in the
data? The default is \code{TRUE} except when \code{.data} has been previously
grouped with \code{.drop = FALSE}. See \code{\link[dplyr:group_by_drop_default]{group_by_drop_default()}} for details.}
}
\value{
A \code{\link[terra:vect]{SpatVector}} object with updated grouping metadata.
}
\description{
Most data operations are done on groups defined by variables.
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}} adds new attributes to an existing
\code{\link[terra:vect]{SpatVector}}
indicating the corresponding groups. See \strong{Methods}.
}
\details{
See \strong{Details} on \code{\link[dplyr:group_by]{dplyr::group_by()}}.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:group_by]{dplyr::group_by()}} method family for
\code{\link[terra:vect]{SpatVector}} objects.
}

\section{Grouping metadata}{


Mixing \CRANpkg{terra} and \CRANpkg{dplyr} syntax on a grouped or row-wise
\code{\link[terra:vect]{SpatVector}}, for example by subsetting with \code{v[1:3, 1:2]},
can corrupt its
grouping metadata. \CRANpkg{tidyterra} attempts to restore this metadata the
next time you use a \CRANpkg{dplyr} verb on the object.

Some operations, such as \code{\link[terra:spatSample]{terra::spatSample()}}, create a new
\code{\link[terra:vect]{SpatVector}}
without preserving grouping metadata. Call \code{\link[=group_by.SpatVector]{group_by.SpatVector()}} or
\code{\link[=rowwise.SpatVector]{rowwise.SpatVector()}} again, as appropriate.
}

\examples{
\donttest{

library(terra)
f <- system.file("ex/lux.shp", package = "terra")
p <- vect(f)

by_name1 <- p |> group_by(NAME_1)

# Grouping does not change how the SpatVector looks.
by_name1

# But it adds metadata for grouping. See the coercion to tibble.

# Not grouped.
p_tbl <- as_tibble(p)
class(p_tbl)
head(p_tbl, 3)

# Grouped.
by_name1_tbl <- as_tibble(by_name1)
class(by_name1_tbl)
head(by_name1_tbl, 3)

# It changes how it acts with the other dplyr verbs:
by_name1 |> summarise(
  pop = mean(POP),
  area = sum(AREA)
)

# Each call to summarise() removes a layer of grouping.
by_name2_name1 <- p |> group_by(NAME_2, NAME_1)

by_name2_name1
group_data(by_name2_name1)

by_name2 <- by_name2_name1 |> summarise(n = dplyr::n())
by_name2
group_data(by_name2)

# To remove grouping, use ungroup().
by_name2 |>
  ungroup() |>
  summarise(n = sum(n))

# By default, group_by() overrides existing grouping.
by_name2_name1 |>
  group_by(ID_1, ID_2) |>
  group_vars()

# Use `.add = TRUE` to append instead.
by_name2_name1 |>
  group_by(ID_1, ID_2, .add = TRUE) |>
  group_vars()

# You can group by expressions. This is shorthand for a mutate() followed
# by a group_by().
p |>
  group_by(ID_COMB = ID_1 * 100 / ID_2) |>
  relocate(ID_COMB, .before = 1)
}
}
\seealso{
\code{\link[dplyr:group_by]{dplyr::group_by()}}, \code{\link[dplyr:ungroup]{dplyr::ungroup()}}.

Other \CRANpkg{dplyr} verbs that operate on groups of rows:
\code{\link[=count.SpatVector]{count.SpatVector()}},
\code{\link[=reframe.SpatVector]{reframe.SpatVector()}},
\code{\link[=rowwise.SpatVector]{rowwise.SpatVector()}},
\code{\link[=summarise.SpatVector]{summarise.SpatVector()}}

Other \CRANpkg{dplyr} grouping methods:
\code{\link[=group_data.SpatVector]{group_data.SpatVector()}},
\code{\link[=group_map.SpatVector]{group_map.SpatVector()}},
\code{\link[=group_nest.SpatVector]{group_nest.SpatVector()}},
\code{\link[=group_split.SpatVector]{group_split.SpatVector()}},
\code{\link[=group_trim.SpatVector]{group_trim.SpatVector()}}
}
\concept{dplyr.group_functions}
\concept{dplyr.groups}
