% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-nest-SpatVector.R
\encoding{UTF-8}
\name{group_nest.SpatVector}
\alias{group_nest.SpatVector}
\alias{nest_by.SpatVector}
\title{Nest grouped \code{\link[terra:vect]{SpatVector}} rows}
\usage{
\method{group_nest}{SpatVector}(.tbl, ..., .key = "data", keep = FALSE)

\method{nest_by}{SpatVector}(.data, ..., .key = "data", .keep = FALSE)
}
\arguments{
\item{.tbl, .data}{A \code{SpatVector}.}

\item{...}{Grouping specification, forwarded to \code{\link[dplyr:group_by]{group_by()}}}

\item{.key}{the name of the list column}

\item{keep, .keep}{Should the grouping columns be kept in the list column.}
}
\value{
A \link[tibble:tbl_df]{tibble} with a list-column of
\code{\link[terra:vect]{SpatVector}}
objects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{group_nest()} and \code{nest_by()} create \link[tibble:tbl_df]{tibble}s with
list-columns containing
\code{\link[terra:vect]{SpatVector}} objects.
}
\section{\CRANpkg{terra} equivalent}{


\code{\link[terra:svc]{terra::svc()}}.
}

\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:group_nest]{dplyr::group_nest()}} method family for
\code{\link[terra:vect]{SpatVector}} objects.

The nested list-column contains \code{\link[terra:vect]{SpatVector}} objects,
preserving the
geometries for each group.
}

\examples{
v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))
v$grp <- rep(c("A", "B"), length.out = nrow(v))

group_nest(v, grp)

nest_by(v, grp)

# Convert to a named SpatVectorCollection.
nested <- group_nest(v, grp)

sv <- pull(nested, data)
names(sv) <- pull(nested, grp)

terra::svc(sv)
}
\seealso{
\code{\link[dplyr:group_nest]{dplyr::group_nest()}}, \code{\link[dplyr:nest_by]{dplyr::nest_by()}}, \code{\link[=nest.SpatVector]{nest.SpatVector()}},
\code{\link[terra:svc]{terra::svc()}}.

Other \CRANpkg{dplyr} grouping methods:
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}},
\code{\link[=group_data.SpatVector]{group_data.SpatVector()}},
\code{\link[=group_map.SpatVector]{group_map.SpatVector()}},
\code{\link[=group_split.SpatVector]{group_split.SpatVector()}},
\code{\link[=group_trim.SpatVector]{group_trim.SpatVector()}}
}
\concept{dplyr.group_functions}
\keyword{internal}
