% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-split-SpatVector.R
\encoding{UTF-8}
\name{group_split.SpatVector}
\alias{group_split.SpatVector}
\title{Split \code{\link[terra:vect]{SpatVector}} by groups}
\usage{
\method{group_split}{SpatVector}(.tbl, ..., .keep = TRUE)
}
\arguments{
\item{.tbl}{A \code{SpatVector} object. See \strong{Methods}.}

\item{...}{If \code{.tbl} is an ungrouped \code{SpatVector}, a grouping specification,
forwarded to \code{\link[=group_by.SpatVector]{group_by.SpatVector()}}.}

\item{.keep}{Should the grouping columns be kept?}
}
\value{
A list of \code{\link[terra:vect]{SpatVector}} objects. Each object contains
the rows
of \code{.tbl} for the associated group and all columns. When \code{.keep = TRUE},
the output includes the grouping variables.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=group_split]{group_split()}} works like \code{\link[base:split]{base::split()}} but:
\itemize{
\item It uses the grouping structure from \code{\link[=group_by.SpatVector]{group_by.SpatVector()}} and therefore
is subject to the data mask.
\item It does not name the elements of the list based on the grouping as this
only works well for a single character grouping variable. Instead,
use \code{\link[=group_keys.SpatVector]{group_keys.SpatVector()}} to access a data frame that defines the
groups.
}

See \code{\link[dplyr:group_split]{dplyr::group_split()}} for more information.
}
\details{
See \strong{Details} on \code{\link[dplyr:group_split]{dplyr::group_split()}}.
}
\section{Lifecycle}{


\code{group_split()} is not stable because you can achieve very similar results
by manipulating the nested column returned from
\code{\link[=nest.SpatVector]{nest(.by =)}}. That also retains the group keys all
within a single data structure. \code{group_split()} may be deprecated in the
future.
}

\section{\CRANpkg{terra} equivalent}{


\code{\link[terra:svc]{terra::svc()}}.
}

\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:group_split]{dplyr::group_split()}} method for
\code{\link[terra:vect]{SpatVector}} objects.
}

\examples{
v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))

v$group <- rep(c("A", "B", "C"), 3)

v |>
  group_by(group) |>
  group_split()

# Coerce the result to a SpatVectorCollection.
v |>
  group_by(group) |>
  group_split() |>
  terra::svc()
}
\seealso{
\code{\link[dplyr:group_split]{dplyr::group_split()}}, \code{\link[terra:svc]{terra::svc()}}.

Other \CRANpkg{dplyr} grouping methods:
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}},
\code{\link[=group_data.SpatVector]{group_data.SpatVector()}},
\code{\link[=group_map.SpatVector]{group_map.SpatVector()}},
\code{\link[=group_nest.SpatVector]{group_nest.SpatVector()}},
\code{\link[=group_trim.SpatVector]{group_trim.SpatVector()}}
}
\concept{dplyr.group_functions}
\keyword{internal}
