% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-trim.SpatVector.R
\encoding{UTF-8}
\name{group_trim.SpatVector}
\alias{group_trim.SpatVector}
\title{Trim grouping structure}
\usage{
\method{group_trim}{SpatVector}(.tbl, .drop = group_by_drop_default(.tbl))
}
\arguments{
\item{.tbl}{A \code{SpatVector} object. See \strong{Methods}.}

\item{.drop}{See \code{\link[=group_by.SpatVector]{group_by.SpatVector()}}.}
}
\value{
A \code{\link[terra:vect]{SpatVector}} object with updated grouping metadata.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This method drops unused levels of all factors that are used as grouping
variables and recalculates the grouping structure.

\code{group_trim()} is particularly useful after a \code{\link[=filter]{filter()}} that is intended
to select a subset of groups.
}
\details{
See \strong{Details} on \code{\link[dplyr:group_trim]{dplyr::group_trim()}}.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:group_trim]{dplyr::group_trim()}} method for
\code{\link[terra:vect]{SpatVector}} objects.
}

\examples{
v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))

v$group <- rep(c("A", "B", "C"), 3)

v |>
  group_by(group) |>
  filter(group == "B", .preserve = TRUE) |>
  group_trim()

}
\seealso{
\code{\link[dplyr:group_trim]{dplyr::group_trim()}}.

Other \CRANpkg{dplyr} grouping methods:
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}},
\code{\link[=group_data.SpatVector]{group_data.SpatVector()}},
\code{\link[=group_map.SpatVector]{group_map.SpatVector()}},
\code{\link[=group_nest.SpatVector]{group_nest.SpatVector()}},
\code{\link[=group_split.SpatVector]{group_split.SpatVector()}}
}
\concept{dplyr.group_functions}
\keyword{internal}
