% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped-SpatVector.R
\encoding{UTF-8}
\name{is_grouped_spatvector}
\alias{is_grouped_spatvector}
\title{A grouped \code{\link[terra:vect]{SpatVector}}}
\usage{
is_grouped_spatvector(x)
}
\arguments{
\item{x}{A \code{SpatVector} created with \code{\link[terra:vect]{terra::vect()}}.}
}
\value{
\code{TRUE} if \code{x} is a grouped \code{\link[terra:vect]{SpatVector}}, otherwise
\code{FALSE}.
}
\description{
The easiest way to create a grouped \code{\link[terra:vect]{SpatVector}} is to call
the
\code{group_by()} method on a \code{SpatVector}: this will take care of capturing
the unevaluated expressions for you. See \code{\link[=group_by.SpatVector]{group_by.SpatVector()}} for
details.

This function is an adapted version of \code{\link[dplyr:is_grouped_df]{dplyr::is_grouped_df()}}.

See also \code{\link[=group_data.SpatVector]{group_data.SpatVector()}} for the helper functions that retrieve
various metadata from a grouped \code{SpatVector}.
}
\examples{
v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))

is_grouped_spatvector(v)

grouped <- group_by(v, iso2)

is_grouped_spatvector(grouped)

}
\concept{helpers}
\keyword{internal}
