% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-regular-grid.R
\encoding{UTF-8}
\name{is_regular_grid}
\alias{is_regular_grid}
\title{Check whether \code{x} and \code{y} positions form a regular grid}
\usage{
is_regular_grid(xy, digits = 6)
}
\arguments{
\item{xy}{A matrix, data frame or tibble with at least two columns
representing \code{x} and \code{y} coordinates.}

\item{digits}{Integer to set the precision for detecting whether points are
on a regular grid (a low number of digits is a low precision).}
}
\value{
\code{NULL}, returned invisibly if the coordinates form a regular grid.
Otherwise, the function raises an error.
}
\description{
Assess whether the \code{x} and \code{y} coordinates of an object form a regular grid.
This function is called for its side effects.

This function is called internally by \code{\link[=as_spatraster]{as_spatraster()}}.
}
\examples{

p <- matrix(1:90, nrow = 45, ncol = 2)

is_regular_grid(p)

# Jitter locations.
set.seed(1234)
jitter <- runif(length(p)) / 10e4
p_jitter <- p + jitter

# Adjust digits.
is_regular_grid(p_jitter, digits = 4)

}
\seealso{
\code{\link[=as_spatraster]{as_spatraster()}}.
}
\concept{helpers}
