% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-Spat.R
\encoding{UTF-8}
\name{mutate.Spat}
\alias{mutate.Spat}
\alias{mutate.SpatRaster}
\alias{mutate.SpatVector}
\title{Create, modify and delete cell values/layers/attributes of \verb{Spat*} objects}
\usage{
\method{mutate}{SpatRaster}(
  .data,
  ...,
  .keep = c("all", "used", "unused", "none"),
  .before = NULL,
  .after = NULL
)

\method{mutate}{SpatVector}(
  .data,
  ...,
  .by = NULL,
  .keep = c("all", "used", "unused", "none"),
  .before = NULL,
  .after = NULL
)
}
\arguments{
\item{.data}{A \code{SpatRaster} created with \code{\link[terra:rast]{terra::rast()}} or a \code{SpatVector}
created with \code{\link[terra:vect]{terra::vect()}}.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Name-value pairs.
The name gives the name of the column in the output.

The value can be:
\itemize{
\item A vector of length 1, which will be recycled to the correct length.
\item A vector the same length as the current group (or the whole data frame
if ungrouped).
\item \code{NULL}, to remove the column.
\item A data frame or tibble, to create multiple columns in the output.
}}

\item{.keep}{Control which columns from \code{.data} are retained in the output. Grouping
columns and columns created by \code{...} are always kept.
\itemize{
\item \code{"all"} retains all columns from \code{.data}. This is the default.
\item \code{"used"} retains only the columns used in \code{...} to create new
columns. This is useful for checking your work, as it displays inputs
and outputs side-by-side.
\item \code{"unused"} retains only the columns \emph{not} used in \code{...} to create new
columns. This is useful if you generate new columns, but no longer need
the columns used to generate them.
\item \code{"none"} doesn't retain any extra columns from \code{.data}. Only the grouping
variables and columns created by \code{...} are kept.
}}

\item{.before, .after}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, control where new columns
should appear (the default is to add to the right hand side). See
\code{\link[dplyr:relocate]{relocate()}} for more details.}

\item{.by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, a selection of columns to
group by for just this operation, functioning as an alternative to \code{\link[dplyr:group_by]{group_by()}}. For
details and examples, see \link[dplyr:dplyr_by]{?dplyr_by}.}
}
\value{
A \verb{Spat*} object of the same class as \code{.data}. See \strong{Methods}.
}
\description{
\code{mutate()} adds new layers/attributes and preserves existing ones on a
\verb{Spat*} object.
}
\section{\CRANpkg{terra} equivalent}{


Some \CRANpkg{terra} methods for modifying cell values:
\code{\link[terra:ifel]{terra::ifel()}}, \code{\link[terra:classify]{terra::classify()}}, \code{\link[terra:clamp]{terra::clamp()}}, \code{\link[terra:app]{terra::app()}},
\code{\link[terra:lapp]{terra::lapp()}} and \code{\link[terra:tapp]{terra::tapp()}}.
}

\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:mutate]{dplyr::mutate()}} methods for \verb{Spat*}
objects.
\subsection{\code{\link[terra:rast]{SpatRaster}}}{

Adds new layers and preserves existing ones. The result is a
\code{\link[terra:rast]{SpatRaster}} with the same extent, resolution and CRS as
\code{.data}. Only the
values and possibly the number of layers are modified.
}

\subsection{\code{\link[terra:vect]{SpatVector}}}{

The result is a \code{\link[terra:vect]{SpatVector}} with the attributes modified and
possibly
renamed as specified in the function call.
}
}

\examples{

library(terra)

# SpatRaster method
f <- system.file("extdata/cyl_temp.tif", package = "tidyterra")
spatrast <- rast(f)

mod <- spatrast |>
  mutate(exp_lyr1 = exp(tavg_04 / 10)) |>
  select(tavg_04, exp_lyr1)

mod
plot(mod)

# SpatVector method
f <- system.file("extdata/cyl.gpkg", package = "tidyterra")
v <- vect(f)

v |>
  mutate(cpro2 = paste0(cpro, "-CyL")) |>
  select(cpro, cpro2)
}
\seealso{
\code{\link[dplyr:mutate]{dplyr::mutate()}} methods.

\CRANpkg{terra} provides several ways to modify \verb{Spat*} objects:
\itemize{
\item \code{\link[terra:ifel]{terra::ifel()}}.
\item \code{\link[terra:classify]{terra::classify()}}.
\item \code{\link[terra:clamp]{terra::clamp()}}.
\item \code{\link[terra:app]{terra::app()}}, \code{\link[terra:lapp]{terra::lapp()}}, \code{\link[terra:tapp]{terra::tapp()}}.
}

Other \CRANpkg{dplyr} verbs that operate on columns:
\code{\link{glimpse.Spat}},
\code{\link{pull.Spat}},
\code{\link{relocate.Spat}},
\code{\link{rename.Spat}},
\code{\link{select.Spat}}
}
\concept{dplyr.cols}
