% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\encoding{UTF-8}
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{A value or the \CRANpkg{magrittr} placeholder.}

\item{rhs}{A function call using \CRANpkg{magrittr} semantics.}
}
\value{
The result of calling \code{rhs(lhs)}.
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\keyword{internal}
