% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-Spat.R
\encoding{UTF-8}
\name{pull.Spat}
\alias{pull.Spat}
\alias{pull.SpatRaster}
\alias{pull.SpatVector}
\title{Extract a single layer/attribute}
\usage{
\method{pull}{SpatRaster}(.data, var = -1, name = NULL, ...)

\method{pull}{SpatVector}(.data, var = -1, name = NULL, ...)
}
\arguments{
\item{.data}{A \code{SpatRaster} created with \code{\link[terra:rast]{terra::rast()}} or a \code{SpatVector}
created with \code{\link[terra:vect]{terra::vect()}}.}

\item{var}{A variable specified as:
\itemize{
\item a literal layer/attribute name.
\item a positive integer, giving the position counting from the left.
\item a negative integer, giving the position counting from the right.
}

The default returns the last layer/attribute (on the assumption that's the
column you've created most recently).

This argument is taken by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names and
column locations).}

\item{name}{An optional parameter that specifies the column to be used
as names for a named vector. Specified in a similar manner as \code{var}.}

\item{...}{Arguments passed to \code{\link[=as_tibble.SpatRaster]{as_tibble.SpatRaster()}} or
\code{\link[=as_tibble.Spat]{as_tibble.Spat()}} methods.}
}
\value{
A vector with one element per cell or geometry in \code{.data}. For
\code{\link[terra:rast]{SpatRaster}} objects, the default (\code{na.rm = FALSE}) keeps
missing values.
Passing \code{na.rm = TRUE} to \code{...} removes cells with missing values and can
shorten the result. See \code{\link[terra:as.data.frame]{terra::as.data.frame()}}.
}
\description{
\code{pull()} is similar to \code{$} on a data frame. It is mostly useful because it
looks nicer in pipes and can optionally name the output.

\strong{You can extract the geographic coordinates of a
\code{\link[terra:rast]{SpatRaster}}}.
Use \code{pull(.data, x, xy = TRUE)}. \code{x} and \code{y} are reserved
names in \CRANpkg{terra}, since they refer to cell coordinates.

See \strong{Examples} and section \strong{Layer and column names} on
\code{\link[=as_tibble.Spat]{as_tibble.Spat()}}.
}
\section{\CRANpkg{terra} equivalent}{


\code{\link[terra:values]{terra::values()}}.
}

\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:pull]{dplyr::pull()}} methods. Each method first
coerces the \verb{Spat*} object to a \link[tibble:tbl_df]{tibble} (see
\link{as_tibble.Spat}) and then
applies \code{\link[dplyr:pull]{dplyr::pull()}} to the tibble.
\subsection{\code{\link[terra:rast]{SpatRaster}}}{

When passing \code{na.rm = TRUE} to \code{...}, only cells with a value other than
\code{NA} are extracted. See \code{\link[terra:as.data.frame]{terra::as.data.frame()}}.

If \code{xy = TRUE} is passed to \code{...}, two columns named \code{x} and \code{y}
(corresponding to the geographic coordinates of each cell) are available
in positions \code{1} and \code{2}. Therefore, \code{pull(.data, 1)} and
\code{pull(.data, 1, xy = TRUE)} return different results.
}

\subsection{\code{\link[terra:vect]{SpatVector}}}{

When passing \code{geom = "WKT"} or \code{geom = "HEX"} to \code{...}, the geometry of the
\code{\link[terra:vect]{SpatVector}} can be extracted with \code{var = geometry}.
Similarly, when using
\code{geom = "XY"}, the coordinates can be extracted with \code{var = x} or
\code{var = y}. See the options in \code{\link[terra:as.data.frame]{terra::as.data.frame()}}.
}
}

\examples{

library(terra)
f <- system.file("extdata/cyl_tile.tif", package = "tidyterra")
r <- rast(f)

# Extract the second layer.
r |>
  pull(2) |>
  head()

# With `xy`, the first two columns are `x` (longitude) and `y` (latitude).

r |>
  pull(2, xy = TRUE) |>
  head()

# With renaming

r |>
  mutate(cat = cut(cyl_tile_3, c(0, 100, 300))) |>
  pull(cyl_tile_3, name = cat) |>
  head()

}
\seealso{
\code{\link[dplyr:pull]{dplyr::pull()}}.

Other \CRANpkg{dplyr} verbs that operate on columns:
\code{\link{glimpse.Spat}},
\code{\link{mutate.Spat}},
\code{\link{relocate.Spat}},
\code{\link{rename.Spat}},
\code{\link{select.Spat}}
}
\concept{dplyr.cols}
