% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate-Spat.R
\encoding{UTF-8}
\name{relocate.Spat}
\alias{relocate.Spat}
\alias{relocate.SpatRaster}
\alias{relocate.SpatVector}
\title{Change layer/attribute order}
\usage{
\method{relocate}{SpatRaster}(.data, ..., .before = NULL, .after = NULL)

\method{relocate}{SpatVector}(.data, ..., .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{A \code{SpatRaster} created with \code{\link[terra:rast]{terra::rast()}} or a \code{SpatVector}
created with \code{\link[terra:vect]{terra::vect()}}.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> layers/attributes to
move.}

\item{.before, .after}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Destination of layers/attributes selected by \code{...}. Supplying neither
will move layers/attributes to the left-hand side. Specifying both is an
error.}
}
\value{
A \verb{Spat*} object of the same class as \code{.data}. See \strong{Methods}.
}
\description{
Use \code{relocate()} to change layer/attribute positions, using the same syntax
as \code{\link{select.Spat}} to make it easy to move blocks of
layers/attributes at once.
}
\section{\CRANpkg{terra} equivalent}{


\code{terra::subset(data, c("name_layer", "name_other_layer"))}.
}

\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:relocate]{dplyr::relocate()}} methods for \verb{Spat*}
objects.
\subsection{\code{\link[terra:rast]{SpatRaster}}}{

Relocates layers of a \code{\link[terra:rast]{SpatRaster}}.
}

\subsection{\code{\link[terra:vect]{SpatVector}}}{

The result is a \code{\link[terra:vect]{SpatVector}} with the attributes in a
different order.
}
}

\examples{

library(terra)

f <- system.file("extdata/cyl_tile.tif", package = "tidyterra")
spatrast <- rast(f) |> mutate(aa = 1, bb = 2, cc = 3)

names(spatrast)

spatrast |>
  relocate(bb, .before = cyl_tile_3) |>
  relocate(cyl_tile_1, .after = last_col())

}
\seealso{
\code{\link[dplyr:relocate]{dplyr::relocate()}}.

Other \CRANpkg{dplyr} verbs that operate on columns:
\code{\link{glimpse.Spat}},
\code{\link{mutate.Spat}},
\code{\link{pull.Spat}},
\code{\link{rename.Spat}},
\code{\link{select.Spat}}
}
\concept{dplyr.cols}
