% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-na-Spat.R
\encoding{UTF-8}
\name{replace_na.Spat}
\alias{replace_na.Spat}
\alias{replace_na.SpatRaster}
\alias{replace_na.SpatVector}
\title{Replace missing values with specified values}
\usage{
\method{replace_na}{SpatRaster}(data, replace = list(), ...)

\method{replace_na}{SpatVector}(data, replace, ...)
}
\arguments{
\item{data}{A \code{SpatRaster} created with \code{\link[terra:rast]{terra::rast()}} or a \code{SpatVector}
created with \code{\link[terra:vect]{terra::vect()}}.}

\item{replace}{A named list of values, with one value for each
layer/attribute that has missing values to be replaced. Each value in
\code{replace} will be cast to the type of the column in \code{data} where it is
used as a replacement.}

\item{...}{Additional arguments for methods. Currently unused.}
}
\value{
A \verb{Spat*} object of the same class as \code{data}. See \strong{Methods}.
}
\description{
Replace \code{NA} values in layers/attributes with specified values.
}
\section{\CRANpkg{terra} equivalent}{


Use \verb{r[is.na(r)] <- <replacement>}.
}

\examples{

library(terra)

f <- system.file("extdata/cyl_temp.tif", package = "tidyterra")
r <- rast(f)

r |> plot()

r |>
  replace_na(list(tavg_04 = 6, tavg_06 = 20)) |>
  plot()

}
\seealso{
\code{\link[tidyr:replace_na]{tidyr::replace_na()}}.

Other \CRANpkg{tidyr} verbs for handling missing values:
\code{\link[=complete.SpatVector]{complete.SpatVector()}},
\code{\link{drop_na.Spat}},
\code{\link[=expand.SpatVector]{expand.SpatVector()}},
\code{\link[=fill.SpatVector]{fill.SpatVector()}}
}
\concept{tidyr.missing}
