% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_fill_topo.R
\name{scale_fill_topo}
\alias{scale_fill_topo}
\alias{scale_fill_topo_c}
\alias{scale_fill_topo_b}
\title{Hypsometric tints fill scales}
\usage{
scale_fill_topo_c(
  ...,
  alpha = 1,
  direction = 1,
  na.value = NA,
  guide = "colourbar"
)

scale_fill_topo_b(
  ...,
  alpha = 1,
  direction = 1,
  na.value = NA,
  guide = "coloursteps"
)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[ggplot2:discrete_scale]{discrete_scale()}},
\code{\link[ggplot2:continuous_scale]{continuous_scale()}}, or \link[ggplot2]{binned_scale} to control name, limits, breaks,
labels and so forth.}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors
are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{na.value}{Missing values will be replaced with this value.}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}
}
\value{
The corresponding ggplot2 layer with the values applied to the
\code{fill} aesthetics.
}
\description{
Implementation of the
\href{https://en.wikipedia.org/wiki/Wikipedia:WikiProject_Maps/Conventions/Topographic_maps}{Wikipedia Colorimetric conventions for topographic maps}.
Two fill scales are provided:
\itemize{
\item \code{scale_fill_topo_c()}: For continuous values.
\item \code{scale_fill_topo_b()}: For binning continuous values.
}
}
\examples{
\donttest{
filepath <- system.file("extdata/cyl_elev.tif", package = "tidyterra")

library(terra)
cyl_elev <- rast(filepath)

library(ggplot2)
ggplot() +
  geom_spatraster(data = cyl_elev) +
  scale_fill_topo_c()

# Binned
ggplot() +
  geom_spatraster(data = cyl_elev) +
  scale_fill_topo_b(breaks = c(100, 250, 500, 1000, 1500, 2000, 2500))
}
}
\seealso{
\code{\link[terra:plot]{terra::plot()}}, \code{\link[ggplot2:scale_viridis]{ggplot2::scale_fill_viridis_c()}}

Other ggplot2 utils:
\code{\link{autoplot}()},
\code{\link{geom_spat_contour}},
\code{\link{geom_spatraster_rgb}()},
\code{\link{geom_spatraster}()},
\code{\link{ggspatvector}},
\code{\link{scale_fill_terrain}}
}
\concept{ggplot2.utils}
