% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-Spat.R
\encoding{UTF-8}
\name{select.Spat}
\alias{select.Spat}
\alias{select.SpatRaster}
\alias{select.SpatVector}
\title{Subset layers/attributes of \verb{Spat*} objects}
\usage{
\method{select}{SpatRaster}(.data, ...)

\method{select}{SpatVector}(.data, ...)
}
\arguments{
\item{.data}{A \code{SpatRaster} created with \code{\link[terra:rast]{terra::rast()}} or a \code{SpatVector}
created with \code{\link[terra:vect]{terra::vect()}}.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more
unquoted expressions separated by commas. Layer/attribute names can be
used as if they were positions in the \verb{Spat*} object, so expressions like
\code{x:y} can be used to select a range of layers/attributes.}
}
\value{
A \verb{Spat*} object of the same class as \code{.data}. See \strong{Methods}.
}
\description{
Select (and optionally rename) attributes/layers in \verb{Spat*} objects, using a
concise mini-language. See \strong{Methods}.
}
\section{\CRANpkg{terra} equivalent}{


\code{\link[terra:subset]{terra::subset()}}.
}

\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:select]{dplyr::select()}} methods for \verb{Spat*}
objects.
\subsection{\code{\link[terra:rast]{SpatRaster}}}{

Selects and renames layers of a \code{\link[terra:rast]{SpatRaster}}. The result is a
\code{SpatRaster}
with the same extent, resolution and CRS as \code{.data}. Only the number and
possibly the names of layers are modified.
}

\subsection{\code{\link[terra:vect]{SpatVector}}}{

The result is a \code{\link[terra:vect]{SpatVector}} with the attributes selected and
possibly
renamed as specified in the function call.
}
}

\examples{

library(terra)

# SpatRaster method

spatrast <- rast(
  crs = "EPSG:3857",
  nrows = 10,
  ncols = 10,
  extent = c(100, 200, 100, 200),
  nlyr = 6,
  vals = seq_len(10 * 10 * 6)
)

spatrast |> select(1)

# By name
spatrast |> select(lyr.1:lyr.4)

# Rename
spatrast |> select(a = lyr.1, c = lyr.6)

# SpatVector method

f <- system.file("extdata/cyl.gpkg", package = "tidyterra")

v <- vect(f)

v

v |> select(1, 3)

v |> select(iso2, name2 = cpro)
}
\seealso{
\code{\link[dplyr:select]{dplyr::select()}}, \code{\link[terra:subset]{terra::subset()}}.

Other \CRANpkg{dplyr} verbs that operate on columns:
\code{\link{glimpse.Spat}},
\code{\link{mutate.Spat}},
\code{\link{pull.Spat}},
\code{\link{relocate.Spat}},
\code{\link{rename.Spat}}
}
\concept{dplyr.cols}
