% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-SpatVector.R
\encoding{UTF-8}
\name{summarise.SpatVector}
\alias{summarise.SpatVector}
\alias{summarize.SpatVector}
\title{Summarize each group of a \code{\link[terra:vect]{SpatVector}} down to one geometry}
\usage{
\method{summarise}{SpatVector}(.data, ..., .by = NULL, .groups = NULL, .dissolve = TRUE)

\method{summarize}{SpatVector}(.data, ..., .by = NULL, .groups = NULL, .dissolve = TRUE)
}
\arguments{
\item{.data}{A \code{SpatVector} created with \code{\link[terra:vect]{terra::vect()}}.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Name-value pairs of
summary functions. The name will be the name of the variable in the result.

The value can be:
\itemize{
\item A vector of length 1, e.g. \code{min(x)}, \code{n()}, or \code{sum(is.na(y))}.
\item A data frame with 1 row, to add multiple columns from a single expression.
}}

\item{.by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, a selection of columns to
group by for just this operation, functioning as an alternative to \code{\link[dplyr:group_by]{group_by()}}. For
details and examples, see \link[dplyr:dplyr_by]{?dplyr_by}.}

\item{.groups}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Grouping structure of the
result.
\itemize{
\item \code{"drop_last"}: drops the last level of grouping. This was the
only supported option before version 1.0.0.
\item \code{"drop"}: All levels of grouping are dropped.
\item \code{"keep"}: Same grouping structure as \code{.data}.
\item \code{"rowwise"}: Each row is its own group.
}

When \code{.groups} is not specified, it is set to \code{"drop_last"} for a grouped
data frame, and \code{"keep"} for a rowwise data frame. In addition, a message
informs you of how the result will be grouped unless the result is
ungrouped, the option \code{"dplyr.summarise.inform"} is set to \code{FALSE}, or when
\code{summarise()} is called from a function in a package.}

\item{.dissolve}{Logical. If \code{TRUE}, dissolve borders between aggregated
geometries.}
}
\value{
A \code{\link[terra:vect]{SpatVector}} object.
}
\description{
\code{summarise()} creates a new \code{\link[terra:vect]{SpatVector}}. It returns one
geometry for each
combination of grouping variables. If there are no grouping variables, the
output will have a single geometry summarizing all observations in the input
and combining all the geometries of the \code{SpatVector}. It will contain one
column for each grouping variable and one column for each of
the summary statistics that you have specified.

\code{summarise.SpatVector()} and \code{summarize.SpatVector()} are synonyms.
}
\section{\CRANpkg{terra} equivalent}{


\code{\link[terra:aggregate]{terra::aggregate()}}.
}

\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:summarise]{dplyr::summarise()}} method for
\code{\link[terra:vect]{SpatVector}} objects.

As in the \CRANpkg{sf} implementation, this function can dissolve geometries
with \code{.dissolve = TRUE} or create \code{MULTI} geometries with
\code{.dissolve = FALSE}. See \strong{Examples}.
}

\examples{
library(terra)
library(ggplot2)

v <- vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))

# Grouped
gr_v <- v |>
  mutate(start_with_s = startsWith(name, "S")) |>
  group_by(start_with_s)

# Dissolve geometries.
diss <- gr_v |>
  summarise(n = dplyr::n(), mean = mean(as.double(cpro)))

diss

autoplot(diss, aes(fill = start_with_s)) +
  ggplot2::labs(title = "Dissolved")

# Keep geometries separate.
no_diss <- gr_v |>
  summarise(n = dplyr::n(), mean = mean(as.double(cpro)), .dissolve = FALSE)

# Same statistic.
no_diss

autoplot(no_diss, aes(fill = start_with_s)) +
  ggplot2::labs(title = "Not Dissolved")
}
\seealso{
\code{\link[dplyr:summarise]{dplyr::summarise()}}, \code{\link[terra:aggregate]{terra::aggregate()}}.

Other \CRANpkg{dplyr} verbs that operate on groups of rows:
\code{\link[=count.SpatVector]{count.SpatVector()}},
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}},
\code{\link[=reframe.SpatVector]{reframe.SpatVector()}},
\code{\link[=rowwise.SpatVector]{rowwise.SpatVector()}}
}
\concept{dplyr.groups}
