% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-Spat.R
\encoding{UTF-8}
\name{tidy.Spat}
\alias{tidy.Spat}
\alias{tidy.SpatRaster}
\alias{tidy.SpatVector}
\alias{tidy.SpatGraticule}
\alias{tidy.SpatExtent}
\title{Tidy \verb{Spat*} objects for plotting}
\usage{
\method{tidy}{SpatRaster}(
  x,
  ...,
  .name_repair = c("unique", "check_unique", "universal", "minimal", "unique_quiet",
    "universal_quiet"),
  maxcell = terra::ncell(x) * 1.1,
  pivot = FALSE
)

\method{tidy}{SpatVector}(x, ...)

\method{tidy}{SpatGraticule}(x, ...)

\method{tidy}{SpatExtent}(x, ..., crs = "")
}
\arguments{
\item{x}{A \code{SpatRaster} created with \code{\link[terra:rast]{terra::rast()}}, a \code{SpatVector}
created with \code{\link[terra:vect]{terra::vect()}}, a \code{SpatGraticule} (see \code{\link[terra:graticule]{terra::graticule()}})
or a \code{SpatExtent} (see \code{\link[terra:ext]{terra::ext()}}).}

\item{...}{Ignored by these methods.}

\item{.name_repair}{Treatment of problematic column names:
\itemize{
\item \code{"minimal"}: No name repair or checks, beyond basic existence,
\item \code{"unique"}: Make sure names are unique and not empty,
\item \code{"check_unique"}: (default value), no name repair, but check they are
\code{unique},
\item \code{"universal"}: Make the names \code{unique} and syntactic
\item \code{"unique_quiet"}: Same as \code{"unique"}, but "quiet"
\item \code{"universal_quiet"}: Same as \code{"universal"}, but "quiet"
\item a function: apply custom name repair (e.g., \code{.name_repair = make.names}
for names in the style of base R).
\item A purrr-style anonymous function, see \code{\link[rlang:as_function]{rlang::as_function()}}
}

This argument is passed on as \code{repair} to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
See there for more details on these terms and the strategies used
to enforce them.}

\item{maxcell}{Positive integer or \code{Inf}. Maximum number of raster cells to
use. Use \code{Inf} to use all cells.}

\item{pivot}{Logical. When \code{TRUE}, a \code{SpatRaster} is returned in \link[tidyr:pivot_longer]{long format}. When \code{FALSE} (the default), it is returned
as a data frame with one column per layer. See \strong{Details}.}

\item{crs}{Input that includes or represents a CRS. It can be an \code{sf} or
\code{sfc} object, a \code{\link[terra:rast]{SpatRaster}} or \code{\link[terra:vect]{SpatVector}}
object, a \code{crs} object from
\code{\link[sf:st_crs]{sf::st_crs()}}, a character string (for example a \href{https://proj.org/en/9.3/operations/projections/index.html}{PROJ string}) or
an integer representing an \href{https://epsg.io/}{EPSG} code.}
}
\value{
\code{\link[=tidy.SpatVector]{tidy.SpatVector()}}, \code{\link[=tidy.SpatGraticule]{tidy.SpatGraticule()}} and \code{\link[=tidy.SpatExtent]{tidy.SpatExtent()}}
return an \code{\link[sf:st_sf]{sf}} object.

\code{\link[=tidy.SpatRaster]{tidy.SpatRaster()}} returns a \link[tibble:tbl_df]{tibble}. See \strong{Methods}.
}
\description{
\code{tidy()} methods for \code{\link[terra:rast]{SpatRaster}},
\code{\link[terra:vect]{SpatVector}},
\code{SpatGraticule} and \code{SpatExtent} objects. These methods return a
\link[tibble:tbl_df]{tibble} for \code{\link[terra:rast]{SpatRaster}} objects and
\code{\link[sf:st_sf]{sf}} objects for vector-based inputs. This interface is similar
to \code{\link{fortify.Spat}} and is provided in case the \code{\link[ggplot2:fortify]{ggplot2::fortify()}} method
is deprecated in the future.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[generics:tidy]{generics::tidy()}} methods for \verb{Spat*}
objects.
\subsection{\code{\link[terra:rast]{SpatRaster}}}{

Returns a \link[tibble:tbl_df]{tibble} that can be used with \verb{ggplot2::geom_*},
such as
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} and \code{\link[ggplot2:geom_raster]{ggplot2::geom_raster()}}.

The resulting \link[tibble:tbl_df]{tibble} includes coordinates in the \code{x} and
\code{y} columns. The
values of each layer are added as extra columns using the layer names from
the \code{\link[terra:rast]{SpatRaster}}.

The CRS of the \code{\link[terra:rast]{SpatRaster}} can be retrieved with
\code{attr(tidySpatRaster, "crs")}.

You can convert the tidy object back to a \code{\link[terra:rast]{SpatRaster}} with
\code{\link[=as_spatraster]{as_spatraster()}}.

When \code{pivot = TRUE}, the \code{\link[terra:rast]{SpatRaster}} is returned in long
format (see
\code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}). The tidy object has the following columns:
\itemize{
\item \code{x}, \code{y}: Coordinates of the cell center in the corresponding CRS.
\item \code{lyr}: Name of the \code{\link[terra:rast]{SpatRaster}} layer associated with
\code{value}.
\item \code{value}: Cell value for the corresponding \code{lyr}.
}

This option can be useful when combining several \verb{geom_*} layers or when
faceting.
}

\subsection{\code{\link[terra:vect]{SpatVector}}, \code{SpatGraticule} and \code{SpatExtent}}{

Returns an \code{\link[sf:st_sf]{sf}} object that can be used with
\code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}}.
}
}

\examples{
\donttest{

# Get a SpatRaster
r <- system.file("extdata/volcano2.tif", package = "tidyterra") |>
  terra::rast() |>
  terra::project("EPSG:4326")

r_tidy <- tidy(r)

r_tidy

# Convert back to a `SpatRaster`.
as_spatraster(r_tidy)

# SpatVector
cyl <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))

cyl

tidy(cyl)

# SpatExtent
ex <- cyl |> terra::ext()

ex

tidy(ex)

# With a CRS
tidy(ex, crs = pull_crs(cyl))

# SpatGraticule
grat <- terra::graticule(60, 30, crs = "+proj=robin")

grat
tidy(grat)
}
}
\seealso{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}}, \code{\link[generics:tidy]{generics::tidy()}}.

Other \CRANpkg{generics} methods:
\code{\link{glance.Spat}},
\code{\link{required_pkgs.Spat}}

Coercing objects:
\code{\link[=as_coordinates]{as_coordinates()}},
\code{\link[=as_sf]{as_sf()}},
\code{\link[=as_spatraster]{as_spatraster()}},
\code{\link[=as_spatvector]{as_spatvector()}},
\code{\link{as_tibble.Spat}},
\code{\link{fortify.Spat}}
}
\concept{coerce}
\concept{generics.methods}
