% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncount-SpatVector.R
\encoding{UTF-8}
\name{uncount.SpatVector}
\alias{uncount.SpatVector}
\title{Duplicate \code{\link[terra:vect]{SpatVector}} rows}
\usage{
\method{uncount}{SpatVector}(data, weights, ..., .remove = TRUE, .id = NULL)
}
\arguments{
\item{data}{A \code{SpatVector}.}

\item{weights}{A vector of weights. Evaluated in the context of \code{data};
supports quasiquotation.}

\item{...}{Additional arguments passed on to methods.}

\item{.remove}{If \code{TRUE}, and \code{weights} is the name of a column in \code{data},
then this column is removed.}

\item{.id}{Supply a string to create a new variable which gives a unique
identifier for each created row.}
}
\value{
A \code{\link[terra:vect]{SpatVector}} object.
}
\description{
\code{uncount()} duplicates rows according to a weighting variable.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[tidyr:uncount]{tidyr::uncount()}} method for
\code{\link[terra:vect]{SpatVector}}
objects.

Each duplicated row keeps the input geometry.
}

\examples{
library(tidyr)

v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))
v$copies <- rep_len(1:2, nrow(v))

uncount(v, copies)

}
\seealso{
\code{\link[tidyr:uncount]{tidyr::uncount()}}.
}
\concept{tidyr.rows}
