% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join-cross.SpatVector.R
\encoding{UTF-8}
\name{cross_join.SpatVector}
\alias{cross_join.SpatVector}
\title{Cross joins for \code{SpatVector} objects}
\usage{
\method{cross_join}{SpatVector}(x, y, ..., copy = FALSE, suffix = c(".x", ".y"))
}
\arguments{
\item{x}{A \code{SpatVector} created with \code{\link[terra:vect]{terra::vect()}}.}

\item{y}{A data frame or other object coercible to a data frame. If a
\code{SpatVector} or \code{sf} object is provided, this method returns an error.}

\item{...}{Additional arguments passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}.}

\item{copy}{If \code{x} and \code{y} are not from the same data source,
and \code{copy} is \code{TRUE}, then \code{y} will be copied into the
same src as \code{x}.  This allows you to join tables across srcs, but
it is a potentially expensive operation so you must opt into it.}

\item{suffix}{If there are non-joined duplicate variables in \code{x} and
\code{y}, these suffixes will be added to the output to disambiguate them.
Should be a character vector of length 2.}
}
\value{
A \code{SpatVector} object.
}
\description{
Cross joins match each row in \code{x} to every row in \code{y}.

See \code{\link[dplyr:cross_join]{dplyr::cross_join()}} for details.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:cross_join]{dplyr::cross_join()}} method.
\subsection{\code{SpatVector}}{

The geometry column has sticky behavior. The result repeats each geometry in
\code{x} once for every row in \code{y}.

If \code{y} has a column named \code{geometry}, it is treated as a regular attribute
and receives the suffix from \code{suffix}.
}
}

\examples{
v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))

labels <- data.frame(period = c("past", "present"))

cross_join(v, labels)
}
\seealso{
\code{\link[dplyr:cross_join]{dplyr::cross_join()}}

Other \CRANpkg{dplyr} verbs that operate on pairs of \code{SpatVector} and data frame objects:
\code{\link{bind_cols.SpatVector}},
\code{\link{bind_rows.SpatVector}},
\code{\link{filter-joins.SpatVector}},
\code{\link{mutate-joins.SpatVector}},
\code{\link[=nest_join.SpatVector]{nest_join.SpatVector()}},
\code{\link{rows.SpatVector}}

Other \CRANpkg{dplyr} methods:
\code{\link[=arrange.SpatVector]{arrange.SpatVector()}},
\code{\link{bind_cols.SpatVector}},
\code{\link{bind_rows.SpatVector}},
\code{\link[=count.SpatVector]{count.SpatVector()}},
\code{\link[=distinct.SpatVector]{distinct.SpatVector()}},
\code{\link{filter-joins.SpatVector}},
\code{\link{filter.Spat}},
\code{\link{glimpse.Spat}},
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}},
\code{\link{mutate-joins.SpatVector}},
\code{\link{mutate.Spat}},
\code{\link[=nest_join.SpatVector]{nest_join.SpatVector()}},
\code{\link{pull.Spat}},
\code{\link[=reframe.SpatVector]{reframe.SpatVector()}},
\code{\link{relocate.Spat}},
\code{\link{rename.Spat}},
\code{\link{rows.SpatVector}},
\code{\link[=rowwise.SpatVector]{rowwise.SpatVector()}},
\code{\link{select.Spat}},
\code{\link{slice.Spat}},
\code{\link[=summarise.SpatVector]{summarise.SpatVector()}}
}
\concept{dplyr.methods}
\concept{dplyr.pairs}
