% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.Spat.R
\name{fortify.Spat}
\alias{fortify.Spat}
\alias{fortify.SpatVector}
\title{Fortify Spat Objects}
\usage{
\method{fortify}{SpatVector}(model, data, ...)
}
\arguments{
\item{model}{A SpatVector \code{\link[terra:vect]{terra::vect()}}.}

\item{data}{Not used by this method.}

\item{...}{other arguments passed to methods}
}
\value{
\code{\link[=fortify.SpatVector]{fortify.SpatVector()}} returns a \code{sf} object. See \strong{Methods}.
}
\description{
Fortify SpatVectors and SpatRasters to data frames for compatibility with
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[ggplot2:fortify]{ggplot2::fortify()}} function.
\subsection{SpatVector}{

Return a \code{sf} object.
}
}

\examples{
\donttest{

# Create a SpatVector
extfile <- system.file("extdata/cyl.gpkg", package = "tidyterra")
cyl <- terra::vect(extfile)

cyl

# To sf
ggplot2::fortify(cyl)

# Now you can use geom_sf()

library(ggplot2)

ggplot(cyl) +
  geom_sf()
}

}
\seealso{
Other ggplot2 utils:
\code{\link{autoplot}()},
\code{\link{geom_spat_contour}},
\code{\link{geom_spatraster_rgb}()},
\code{\link{geom_spatraster}()},
\code{\link{ggspatvector}},
\code{\link{stat_spat_coordinates}()}
}
\concept{ggplot2.methods}
\concept{ggplot2.utils}
