% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_spatraster_rgb.R
\name{geom_spatraster_rgb}
\alias{geom_spatraster_rgb}
\title{Visualise SpatRaster objects as images}
\source{
Based on the \code{layer_spatial()} implementation on ggspatial package.
Thanks to \href{https://github.com/paleolimbot}{Dewey Dunnington} and
\href{https://github.com/paleolimbot/ggspatial/graphs/contributors}{ggspatial contributors}.
}
\usage{
geom_spatraster_rgb(
  mapping = aes(),
  data,
  interpolate = TRUE,
  r = 1,
  g = 2,
  b = 3,
  alpha = 0.9,
  maxcell = 5e+05,
  ...
)
}
\arguments{
\item{mapping}{Ignored.}

\item{data}{A SpatRaster object.}

\item{interpolate}{If \code{TRUE} interpolate linearly, if \code{FALSE}
(the default) don't interpolate.}

\item{r, g, b}{Integer representing the number of layer of \code{data} to be
considered as the red (\code{r}), green (\code{g}) and blue (\code{b}) channel.}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{maxcell}{positive integer. Maximum number of cells to use for
the plot.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
A ggplot2 layer
}
\description{
This geom is used to visualise SpatRaster objects (see \code{\link[terra:rast]{terra::rast()}}) as
RGB images. The layers are combined such that they represent the red,
green and blue channel.

For plotting SpatRaster objects by layer values use \code{\link[=geom_spatraster]{geom_spatraster()}}.

The underlying implementation is based on \code{\link[ggplot2:geom_tile]{ggplot2::geom_raster()}}.
}
\section{terra equivalent}{


\code{\link[terra:plotRGB]{terra::plotRGB()}}
}

\section{Coords}{


When the SpatRaster does not present a crs (i.e.,
\code{terra::crs(rast) == ""}) the geom does not make any assumption on the
scales.

On SpatRaster that have a crs, the geom uses \code{\link[ggplot2:ggsf]{ggplot2::coord_sf()}} to adjust
the scales. That means that also the SpatRaster may be reprojected.
}

\section{Aesthetics}{


No \code{aes()} is required. In fact, \code{aes()} will be ignored.
}

\examples{
\donttest{

# Tile of Castille and Leon (Spain) from OpenStreetMap
file_path <- system.file("extdata/cyl_tile.tif", package = "tidyterra")

library(terra)
tile <- rast(file_path)

library(ggplot2)


ggplot() +
  geom_spatraster_rgb(data = tile) +
  # You can use coord_sf
  coord_sf(crs = 3035)

# Combine with sf objects
vect_path <- system.file("extdata/cyl.gpkg", package = "tidyterra")

cyl_sf <- sf::st_read(vect_path)

ggplot(cyl_sf) +
  geom_spatraster_rgb(data = tile) +
  geom_sf(aes(fill = iso2)) +
  coord_sf(crs = 3857) +
  scale_fill_viridis_d(alpha = 0.7)
}
}
\seealso{
\code{\link[ggplot2:geom_tile]{ggplot2::geom_raster()}}, \code{\link[ggplot2:ggsf]{ggplot2::coord_sf()}}. You can get
also RGB tiles from the {maptiles} package, see \code{\link[maptiles:get_tiles]{maptiles::get_tiles()}}.

Other ggplot2 utils:
\code{\link{autoplot}()},
\code{\link{geom_spatraster}()},
\code{\link{ggspatvector}},
\code{\link{scale_fill_terrain}}
}
\concept{ggplot2.utils}
