% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped-SpatVector.R
\encoding{UTF-8}
\name{is_grouped_spatvector}
\alias{is_grouped_spatvector}
\title{A grouped \code{SpatVector}}
\usage{
is_grouped_spatvector(x)
}
\arguments{
\item{x}{A \code{SpatVector} created with \code{\link[terra:vect]{terra::vect()}}.}
}
\description{
The easiest way to create a grouped \code{SpatVector} is to call the
\code{group_by()} method on a \code{SpatVector}: this will take care of capturing
the unevaluated expressions for you. See \code{\link[=group_by.SpatVector]{group_by.SpatVector()}} for details.

This function is the adapted version of \code{\link[dplyr:is_grouped_df]{dplyr::is_grouped_df()}}.

See also \code{\link[=group_data.SpatVector]{group_data.SpatVector()}} for the accessory functions that retrieve
various metadata from a grouped \code{SpatVector}.
}
\seealso{
Other helpers:
\code{\link[=compare_spatrasters]{compare_spatrasters()}},
\code{\link[=is_regular_grid]{is_regular_grid()}},
\code{\link[=pull_crs]{pull_crs()}}
}
\concept{helpers}
\keyword{internal}
