% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped-SpatVector.R
\name{is_grouped_spatvector}
\alias{is_grouped_spatvector}
\title{A grouped SpatVector.}
\usage{
is_grouped_spatvector(x)
}
\arguments{
\item{x}{a SpatVector.}
}
\description{
The easiest way to create a grouped SpatVector is to call the
\code{group_by()} method on a Spatvector: this will take care of capturing
the unevaluated expressions for you. See \code{\link[=group_by.SpatVector]{group_by.SpatVector()}} for details.

This function is the adapted version of \code{\link[dplyr:grouped_df]{dplyr::is_grouped_df()}}.

See also \code{\link[=group_data.SpatVector]{group_data.SpatVector()}} for the accessor functions that retrieve
various metadata from a grouped SpatVector.
}
\seealso{
Other helpers: 
\code{\link{coltab_palette}()},
\code{\link{compare_spatrasters}()},
\code{\link{is_regular_grid}()},
\code{\link{pull_crs}()}
}
\concept{helpers}
\keyword{internal}
