% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest-SpatVector.R
\encoding{UTF-8}
\name{nest.SpatVector}
\alias{nest.SpatVector}
\title{Nest \code{SpatVector} rows}
\usage{
\method{nest}{SpatVector}(.data, ..., .by = NULL, .key = NULL, .names_sep = NULL)
}
\arguments{
\item{.data}{A \code{SpatVector}.}

\item{...}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> Columns to nest; these will
appear in the inner data frames.

Specified using name-variable pairs of the form
\code{new_col = c(col1, col2, col3)}. The right hand side can be any valid
tidyselect expression.

If not supplied, then \code{...} is derived as all columns \emph{not} selected by
\code{.by}, and will use the column name from \code{.key}.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}:
previously you could write \code{df |> nest(x, y, z)}.
Convert to \code{df |> nest(data = c(x, y, z))}.}

\item{.by}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> Columns to nest \emph{by}; these
will remain in the outer data frame.

\code{.by} can be used in place of or in conjunction with columns supplied
through \code{...}.

If not supplied, then \code{.by} is derived as all columns \emph{not} selected by
\code{...}.}

\item{.key}{The name of the resulting nested column. Only applicable when
\code{...} isn't specified, i.e. in the case of \code{df |> nest(.by = x)}.

If \code{NULL}, then \code{"data"} will be used by default.}

\item{.names_sep}{If \code{NULL}, the default, the inner names will come from
the former outer names. If a string, the  new inner names will use the
outer names with \code{names_sep} automatically stripped. This makes
\code{names_sep} roughly symmetric between nesting and unnesting.}
}
\value{
A tibble with one or more list-columns of \code{SpatVector} objects.
}
\description{
\code{nest()} creates list-columns of \code{SpatVector} objects.
}
\section{\CRANpkg{terra} equivalent}{


\code{\link[terra:svc]{terra::svc()}}.
}

\section{Methods}{


Implementation of the \strong{generic} \code{\link[tidyr:nest]{tidyr::nest()}} method.
\subsection{\code{SpatVector}}{

The geometry column must be nested with the other attributes that form each
nested \code{SpatVector}. These nested list-columns contain \code{SpatVector} objects
and cannot be passed directly to \code{\link[tidyr:unnest]{tidyr::unnest()}}.
}
}

\examples{
v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))

v |>
  group_by(cpro) |>
  nest()

# Convert to a named SpatVectorCollection.
nested <- nest(v, .by = cpro)

sv <- pull(nested, data)
names(sv) <- pull(nested, cpro)

terra::svc(sv)

}
\seealso{
\code{\link[tidyr:nest]{tidyr::nest()}}, \code{\link[terra:svc]{terra::svc()}}

Other \CRANpkg{tidyr} methods:
\code{\link[=complete.SpatVector]{complete.SpatVector()}},
\code{\link{drop_na.Spat}},
\code{\link[=expand.SpatVector]{expand.SpatVector()}},
\code{\link[=fill.SpatVector]{fill.SpatVector()}},
\code{\link[=pivot_longer.SpatVector]{pivot_longer.SpatVector()}},
\code{\link[=pivot_wider.SpatVector]{pivot_wider.SpatVector()}},
\code{\link{replace_na.Spat}},
\code{\link[=uncount.SpatVector]{uncount.SpatVector()}},
\code{\link{unite.Spat}}
}
\concept{tidyr.methods}
\concept{tidyr.nest}
