% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest-join-SpatVector.R
\encoding{UTF-8}
\name{nest_join.SpatVector}
\alias{nest_join.SpatVector}
\title{Nest join \code{SpatVector} objects}
\usage{
\method{nest_join}{SpatVector}(
  x,
  y,
  by = NULL,
  copy = FALSE,
  keep = NULL,
  name = NULL,
  ...,
  na_matches = c("na", "never")
)
}
\arguments{
\item{x}{A \code{SpatVector}.}

\item{y}{A data frame. Spatial \code{y} inputs are not supported, use spatial
joins from \CRANpkg{terra} for that workflow.}

\item{by}{A join specification created with \code{\link[dplyr:join_by]{join_by()}}, or a character
vector of variables to join by.

If \code{NULL}, the default, \verb{*_join()} will perform a natural join, using all
variables in common across \code{x} and \code{y}. A message lists the variables so
that you can check they're correct; suppress the message by supplying \code{by}
explicitly.

To join on different variables between \code{x} and \code{y}, use a \code{\link[dplyr:join_by]{join_by()}}
specification. For example, \code{join_by(a == b)} will match \code{x$a} to \code{y$b}.

To join by multiple variables, use a \code{\link[dplyr:join_by]{join_by()}} specification with
multiple expressions. For example, \code{join_by(a == b, c == d)} will match
\code{x$a} to \code{y$b} and \code{x$c} to \code{y$d}. If the column names are the same between
\code{x} and \code{y}, you can shorten this by listing only the variable names, like
\code{join_by(a, c)}.

\code{\link[dplyr:join_by]{join_by()}} can also be used to perform inequality, rolling, and overlap
joins. See the documentation at \link[dplyr:join_by]{?join_by} for details on
these types of joins.

For simple equality joins, you can alternatively specify a character vector
of variable names to join by. For example, \code{by = c("a", "b")} joins \code{x$a}
to \code{y$a} and \code{x$b} to \code{y$b}. If variable names differ between \code{x} and \code{y},
use a named character vector like \code{by = c("x_a" = "y_a", "x_b" = "y_b")}.

To perform a cross-join, generating all combinations of \code{x} and \code{y}, see
\code{\link[dplyr:cross_join]{cross_join()}}.}

\item{copy}{If \code{x} and \code{y} are not from the same data source,
and \code{copy} is \code{TRUE}, then \code{y} will be copied into the
same src as \code{x}.  This allows you to join tables across srcs, but
it is a potentially expensive operation so you must opt into it.}

\item{keep}{Should the new list-column contain join keys? The default
will preserve the join keys for inequality joins.}

\item{name}{The name of the list-column created by the join. If \code{NULL},
the default, the name of \code{y} is used.}

\item{...}{Other parameters passed onto methods.}

\item{na_matches}{Should two \code{NA} or two \code{NaN} values match?
\itemize{
\item \code{"na"}, the default, treats two \code{NA} or two \code{NaN} values as equal, like
\code{\%in\%}, \code{\link[=match]{match()}}, and \code{\link[=merge]{merge()}}.
\item \code{"never"} treats two \code{NA} or two \code{NaN} values as different, and will
never match them together or to any other values. This is similar to joins
for database sources and to \code{base::merge(incomparables = NA)}.
}}
}
\value{
A \link[tibble:tbl_df]{tibble}.
}
\description{
\code{nest_join()} returns a tibble with the attributes and geometry of \code{x}, plus
a list-column containing matching rows from \code{y}.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:nest_join]{dplyr::nest_join()}} method.
\subsection{\code{SpatVector}}{

The output is a tibble with the attributes and WKT geometry of \code{x}, plus a
list-column with matching rows from \code{y}.
}
}

\examples{
v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))
extra <- tibble::tibble(cpro = c("05", "09"), value = c(1, 2))

nest_join(v, extra, by = "cpro")
}
\seealso{
\code{\link[dplyr:nest_join]{dplyr::nest_join()}}

Other \CRANpkg{dplyr} verbs that operate on pairs of \code{SpatVector} and data frame objects:
\code{\link{bind_cols.SpatVector}},
\code{\link{bind_rows.SpatVector}},
\code{\link[=cross_join.SpatVector]{cross_join.SpatVector()}},
\code{\link{filter-joins.SpatVector}},
\code{\link{mutate-joins.SpatVector}},
\code{\link{rows.SpatVector}}

Other \CRANpkg{dplyr} methods:
\code{\link[=arrange.SpatVector]{arrange.SpatVector()}},
\code{\link{bind_cols.SpatVector}},
\code{\link{bind_rows.SpatVector}},
\code{\link[=count.SpatVector]{count.SpatVector()}},
\code{\link[=cross_join.SpatVector]{cross_join.SpatVector()}},
\code{\link[=distinct.SpatVector]{distinct.SpatVector()}},
\code{\link{filter-joins.SpatVector}},
\code{\link{filter.Spat}},
\code{\link{glimpse.Spat}},
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}},
\code{\link{mutate-joins.SpatVector}},
\code{\link{mutate.Spat}},
\code{\link{pull.Spat}},
\code{\link[=reframe.SpatVector]{reframe.SpatVector()}},
\code{\link{relocate.Spat}},
\code{\link{rename.Spat}},
\code{\link{rows.SpatVector}},
\code{\link[=rowwise.SpatVector]{rowwise.SpatVector()}},
\code{\link{select.Spat}},
\code{\link{slice.Spat}},
\code{\link[=summarise.SpatVector]{summarise.SpatVector()}}
}
\concept{dplyr.methods}
\concept{dplyr.pairs}
