% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange-SpatVector.R, R/as-tibble-Spat.R,
%   R/autoplot-Spat.R, R/count-tally-SpatVector.R, R/distinct-SpatVector.R,
%   R/drop-na-Spat.R, R/fill-SpatVector.R, R/filter-Spat.R, R/fortify-Spat.R,
%   R/glance-Spat.R, R/glimpse-Spat.R, R/group-by-SpatVector.R,
%   R/group-data-SpatVector.R, R/join-SpatVector.R, R/mutate-Spat.R,
%   R/pivot-long-SpatVector.R, R/pivot-wide-SpatVector.R, R/pull-Spat.R,
%   R/relocate-Spat.R, R/rename-Spat.R, R/replace-na-Spat.R,
%   R/required-pkgs-Spat.R, R/rowwise-SpatVector.R, R/select-Spat.R,
%   R/slice-Spat.R, R/summarise-SpatVector.R, R/tidy-Spat.R,
%   R/tidyterra-package.R, R/transmute-Spat.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{as_tibble}
\alias{autoplot}
\alias{count}
\alias{tally}
\alias{add_count}
\alias{distinct}
\alias{drop_na}
\alias{fill}
\alias{filter}
\alias{filter_out}
\alias{fortify}
\alias{glance}
\alias{glimpse}
\alias{group_by}
\alias{ungroup}
\alias{group_by_drop_default}
\alias{group_data}
\alias{group_keys}
\alias{group_rows}
\alias{group_indices}
\alias{group_vars}
\alias{groups}
\alias{group_size}
\alias{n_groups}
\alias{inner_join}
\alias{left_join}
\alias{right_join}
\alias{full_join}
\alias{semi_join}
\alias{anti_join}
\alias{mutate}
\alias{pivot_longer}
\alias{pivot_wider}
\alias{pull}
\alias{relocate}
\alias{rename}
\alias{rename_with}
\alias{replace_na}
\alias{required_pkgs}
\alias{rowwise}
\alias{select}
\alias{slice}
\alias{slice_head}
\alias{slice_max}
\alias{slice_min}
\alias{slice_tail}
\alias{slice_sample}
\alias{summarise}
\alias{summarize}
\alias{tidy}
\alias{aes}
\alias{after_stat}
\alias{transmute}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:add_count]{add_count()}}, \code{\link[dplyr:anti_join]{anti_join()}}, \code{\link[dplyr:arrange]{arrange()}}, \code{\link[dplyr:count]{count()}}, \code{\link[dplyr:distinct]{distinct()}}, \code{\link[dplyr:filter]{filter()}}, \code{\link[dplyr:filter_out]{filter_out()}}, \code{\link[dplyr:full_join]{full_join()}}, \code{\link[dplyr:glimpse]{glimpse}}, \code{\link[dplyr:group_by]{group_by()}}, \code{\link[dplyr:group_by_drop_default]{group_by_drop_default()}}, \code{\link[dplyr:group_data]{group_data()}}, \code{\link[dplyr:group_indices]{group_indices()}}, \code{\link[dplyr:group_keys]{group_keys()}}, \code{\link[dplyr:group_rows]{group_rows()}}, \code{\link[dplyr:group_size]{group_size()}}, \code{\link[dplyr:group_vars]{group_vars()}}, \code{\link[dplyr:groups]{groups()}}, \code{\link[dplyr:groups]{groups()}}, \code{\link[dplyr:inner_join]{inner_join()}}, \code{\link[dplyr:left_join]{left_join()}}, \code{\link[dplyr:mutate]{mutate()}}, \code{\link[dplyr:n_groups]{n_groups()}}, \code{\link[dplyr:pull]{pull()}}, \code{\link[dplyr:relocate]{relocate()}}, \code{\link[dplyr:rename]{rename()}}, \code{\link[dplyr:rename_with]{rename_with()}}, \code{\link[dplyr:right_join]{right_join()}}, \code{\link[dplyr:rowwise]{rowwise()}}, \code{\link[dplyr:select]{select()}}, \code{\link[dplyr:semi_join]{semi_join()}}, \code{\link[dplyr:slice]{slice()}}, \code{\link[dplyr:slice_head]{slice_head()}}, \code{\link[dplyr:slice_max]{slice_max()}}, \code{\link[dplyr:slice_min]{slice_min()}}, \code{\link[dplyr:slice_sample]{slice_sample()}}, \code{\link[dplyr:slice_tail]{slice_tail()}}, \code{\link[dplyr:summarise]{summarise()}}, \code{\link[dplyr:summarize]{summarize()}}, \code{\link[dplyr:tally]{tally()}}, \code{\link[dplyr:transmute]{transmute()}}, \code{\link[dplyr:ungroup]{ungroup()}}}

  \item{generics}{\code{\link[generics:glance]{glance()}}, \code{\link[generics:required_pkgs]{required_pkgs()}}, \code{\link[generics:tidy]{tidy()}}}

  \item{ggplot2}{\code{\link[ggplot2:aes]{aes()}}, \code{\link[ggplot2:after_stat]{after_stat()}}, \code{\link[ggplot2:autoplot]{autoplot()}}, \code{\link[ggplot2:fortify]{fortify()}}}

  \item{tibble}{\code{\link[tibble:as_tibble]{as_tibble()}}}

  \item{tidyr}{\code{\link[tidyr:drop_na]{drop_na()}}, \code{\link[tidyr:fill]{fill()}}, \code{\link[tidyr:pivot_longer]{pivot_longer()}}, \code{\link[tidyr:pivot_wider]{pivot_wider()}}, \code{\link[tidyr:replace_na]{replace_na()}}}
}}

