% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.Spat.R, R/as_tibble.Spat.R,
%   R/autoplot.Spat.R, R/distinct.Spat.R, R/drop_na.Spat.R, R/filter.Spat.R,
%   R/fortify.Spat.R, R/glimpse.Spat.R, R/mutate.Spat.R, R/pull.Spat.R,
%   R/relocate.Spat.R, R/rename.Spat.R, R/replace_na.Spat.R, R/select.Spat.R,
%   R/slice.Spat.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{as_tibble}
\alias{autoplot}
\alias{distinct}
\alias{drop_na}
\alias{filter}
\alias{fortify}
\alias{glimpse}
\alias{mutate}
\alias{transmute}
\alias{pull}
\alias{relocate}
\alias{rename}
\alias{rename_with}
\alias{replace_na}
\alias{select}
\alias{slice}
\alias{slice_head}
\alias{slice_max}
\alias{slice_min}
\alias{slice_tail}
\alias{slice_sample}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{glimpse}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{relocate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:rename]{rename_with}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr:slice]{slice_head}}, \code{\link[dplyr:slice]{slice_max}}, \code{\link[dplyr:slice]{slice_min}}, \code{\link[dplyr:slice]{slice_sample}}, \code{\link[dplyr:slice]{slice_tail}}, \code{\link[dplyr]{transmute}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}, \code{\link[ggplot2]{fortify}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{tidyr}{\code{\link[tidyr]{drop_na}}, \code{\link[tidyr]{replace_na}}}
}}

