% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange-SpatVector.R, R/as_tibble-Spat.R,
%   R/autoplot-Spat.R, R/count-tally-SpatVector.R, R/distinct-SpatVector.R,
%   R/drop_na-Spat.R, R/filter-Spat.R, R/fortify-Spat.R, R/glimpse-Spat.R,
%   R/group-by-SpatVector.R, R/group-data-SpatVector.R, R/join-SpatVector.R,
%   R/mutate-Spat.R, R/pull-Spat.R, R/relocate-Spat.R, R/rename-Spat.R,
%   R/replace_na-Spat.R, R/rowwise-SpatVector.R, R/select-Spat.R,
%   R/slice-Spat.R, R/summarise-SpatVector.R, R/tidyterra-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{as_tibble}
\alias{autoplot}
\alias{count}
\alias{tally}
\alias{distinct}
\alias{drop_na}
\alias{filter}
\alias{fortify}
\alias{glimpse}
\alias{group_by}
\alias{ungroup}
\alias{group_by_drop_default}
\alias{group_data}
\alias{group_keys}
\alias{group_rows}
\alias{group_indices}
\alias{group_vars}
\alias{groups}
\alias{group_size}
\alias{n_groups}
\alias{inner_join}
\alias{left_join}
\alias{right_join}
\alias{full_join}
\alias{semi_join}
\alias{anti_join}
\alias{mutate}
\alias{transmute}
\alias{pull}
\alias{relocate}
\alias{rename}
\alias{rename_with}
\alias{replace_na}
\alias{rowwise}
\alias{select}
\alias{slice}
\alias{slice_head}
\alias{slice_max}
\alias{slice_min}
\alias{slice_tail}
\alias{slice_sample}
\alias{summarise}
\alias{summarize}
\alias{aes}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:filter-joins]{anti_join}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr:mutate-joins]{full_join}}, \code{\link[dplyr]{glimpse}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{group_by_drop_default}}, \code{\link[dplyr]{group_data}}, \code{\link[dplyr:group_data]{group_indices}}, \code{\link[dplyr:group_data]{group_keys}}, \code{\link[dplyr:group_data]{group_rows}}, \code{\link[dplyr:group_data]{group_size}}, \code{\link[dplyr:group_data]{group_vars}}, \code{\link[dplyr:group_data]{groups}}, \code{\link[dplyr:group_data]{groups}}, \code{\link[dplyr:mutate-joins]{inner_join}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr:group_data]{n_groups}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{relocate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:rename]{rename_with}}, \code{\link[dplyr:mutate-joins]{right_join}}, \code{\link[dplyr]{rowwise}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:filter-joins]{semi_join}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr:slice]{slice_head}}, \code{\link[dplyr:slice]{slice_max}}, \code{\link[dplyr:slice]{slice_min}}, \code{\link[dplyr:slice]{slice_sample}}, \code{\link[dplyr:slice]{slice_tail}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr:summarise]{summarize}}, \code{\link[dplyr:count]{tally}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr:group_by]{ungroup}}}

  \item{ggplot2}{\code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{autoplot}}, \code{\link[ggplot2]{fortify}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{tidyr}{\code{\link[tidyr]{drop_na}}, \code{\link[tidyr]{replace_na}}}
}}

