% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reframe-SpatVector.R
\encoding{UTF-8}
\name{reframe.SpatVector}
\alias{reframe.SpatVector}
\title{Reframe each group of a \code{SpatVector}}
\usage{
\method{reframe}{SpatVector}(.data, ..., .by = NULL, .dissolve = TRUE)
}
\arguments{
\item{.data}{A \code{SpatVector} created with \code{\link[terra:vect]{terra::vect()}}.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Name-value pairs of
summary functions. The name will be the name of the variable in the result.

The value can be:
\itemize{
\item A vector of length 1, e.g. \code{min(x)}, \code{n()}, or \code{sum(is.na(y))}.
\item A data frame with 1 row, to add multiple columns from a single expression.
}}

\item{.by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, a selection of columns to
group by for just this operation, functioning as an alternative to \code{\link[dplyr:group_by]{group_by()}}. For
details and examples, see \link[dplyr:dplyr_by]{?dplyr_by}.}

\item{.dissolve}{Logical. If \code{TRUE}, dissolve borders between aggregated
geometries.}
}
\value{
A \code{SpatVector}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{reframe()} can return any number of rows per group. The geometry of each
group is aggregated and repeated for each row created for that group.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:reframe]{dplyr::reframe()}} method for \code{SpatVector}
objects.

For grouped inputs, and for calls using \code{.by}, geometries are aggregated per
group. If a group produces more than one row, the aggregated group geometry
is repeated for each output row.
}

\examples{
v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))
v$grp <- rep(c("A", "B"), length.out = nrow(v))

v |>
  reframe(value = c(min(as.double(cpro)), max(as.double(cpro))), .by = grp)

v |>
  rowwise() |>
  reframe(value = 1:2)
}
\seealso{
\code{\link[dplyr:reframe]{dplyr::reframe()}}.

Other \CRANpkg{dplyr} verbs that operate on groups of rows:
\code{\link[=count.SpatVector]{count.SpatVector()}},
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}},
\code{\link[=rowwise.SpatVector]{rowwise.SpatVector()}},
\code{\link[=summarise.SpatVector]{summarise.SpatVector()}}
}
\concept{dplyr.groups}
