% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reframe-SpatVector.R
\encoding{UTF-8}
\name{reframe.SpatVector}
\alias{reframe.SpatVector}
\title{Reframe each group of a \code{SpatVector}}
\usage{
\method{reframe}{SpatVector}(.data, ..., .by = NULL, .dissolve = TRUE)
}
\arguments{
\item{.data}{A \code{SpatVector}.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}>

Name-value pairs of functions. The name will be the name of the variable in
the result. The value can be a vector of any length.

Unnamed data frame values add multiple columns from a single expression.}

\item{.by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, a selection of columns to
group by for just this operation, functioning as an alternative to \code{\link[dplyr:group_by]{group_by()}}. For
details and examples, see \link[dplyr:dplyr_by]{?dplyr_by}.}

\item{.dissolve}{Logical. If \code{TRUE}, dissolve borders between aggregated
geometries.}
}
\value{
A \code{SpatVector}.
}
\description{
\code{reframe()} can return any number of rows per group. The geometry of each
group is aggregated and repeated for each row created for that group.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:reframe]{dplyr::reframe()}} method.
\subsection{\code{SpatVector}}{

For grouped inputs, and for calls using \code{.by}, geometries are aggregated per
group. If a group produces more than one row, the aggregated group geometry
is repeated for each output row.
}
}

\examples{
v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))
v$grp <- rep(c("A", "B"), length.out = nrow(v))

v |>
  reframe(value = c(min(as.double(cpro)), max(as.double(cpro))), .by = grp)

v |>
  rowwise() |>
  reframe(value = 1:2)
}
\seealso{
\code{\link[dplyr:reframe]{dplyr::reframe()}}, \code{\link[=summarise.SpatVector]{summarise.SpatVector()}}

Other \CRANpkg{dplyr} single-table verbs:
\code{\link[=arrange.SpatVector]{arrange.SpatVector()}},
\code{\link{filter.Spat}},
\code{\link{mutate.Spat}},
\code{\link{rename.Spat}},
\code{\link{select.Spat}},
\code{\link{slice.Spat}},
\code{\link[=summarise.SpatVector]{summarise.SpatVector()}}

Other \CRANpkg{dplyr} verbs that operate on groups of rows:
\code{\link[=count.SpatVector]{count.SpatVector()}},
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}},
\code{\link[=rowwise.SpatVector]{rowwise.SpatVector()}},
\code{\link[=summarise.SpatVector]{summarise.SpatVector()}}

Other \CRANpkg{dplyr} methods:
\code{\link[=arrange.SpatVector]{arrange.SpatVector()}},
\code{\link{bind_cols.SpatVector}},
\code{\link{bind_rows.SpatVector}},
\code{\link[=count.SpatVector]{count.SpatVector()}},
\code{\link[=cross_join.SpatVector]{cross_join.SpatVector()}},
\code{\link[=distinct.SpatVector]{distinct.SpatVector()}},
\code{\link{filter-joins.SpatVector}},
\code{\link{filter.Spat}},
\code{\link{glimpse.Spat}},
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}},
\code{\link{mutate-joins.SpatVector}},
\code{\link{mutate.Spat}},
\code{\link[=nest_join.SpatVector]{nest_join.SpatVector()}},
\code{\link{pull.Spat}},
\code{\link{relocate.Spat}},
\code{\link{rename.Spat}},
\code{\link{rows.SpatVector}},
\code{\link[=rowwise.SpatVector]{rowwise.SpatVector()}},
\code{\link{select.Spat}},
\code{\link{slice.Spat}},
\code{\link[=summarise.SpatVector]{summarise.SpatVector()}}
}
\concept{dplyr.groups}
\concept{dplyr.methods}
\concept{dplyr.single_table}
