% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-Spat.R
\name{rename.Spat}
\alias{rename.Spat}
\alias{rename.SpatRaster}
\alias{rename_with.SpatRaster}
\alias{rename.SpatVector}
\alias{rename_with.SpatVector}
\title{Rename layers/attributes}
\usage{
\method{rename}{SpatRaster}(.data, ...)

\method{rename_with}{SpatRaster}(.data, .fn, .cols = everything(), ...)

\method{rename}{SpatVector}(.data, ...)

\method{rename_with}{SpatVector}(.data, .fn, .cols = everything(), ...)
}
\arguments{
\item{.data}{A \code{SpatRaster} created with \code{\link[terra:rast]{terra::rast()}} or a \code{SpatVector}
created with \code{\link[terra:vect]{terra::vect()}}.}

\item{...}{For \verb{rename.Spat*()}: <\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Use \code{new_name = old_name} to rename selected variables.

For \verb{rename_with.Spat*()}: additional arguments passed onto \code{.fn}.}

\item{.fn}{A function used to transform the selected \code{.cols}. Should
return a character vector the same length as the input.}

\item{.cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to rename;
defaults to all columns.}
}
\value{
A \verb{Spat*} object of the same class than \code{.data}. See \strong{Methods}.
}
\description{
\code{rename()} changes the names of individual layers/attributes using
\code{new_name = old_name} syntax; \code{rename_with()} renames layers/attributes
using a function.
}
\section{\CRANpkg{terra} equivalent}{


\verb{names(Spat*) <- c("a", "b", "c")}
}

\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:rename]{dplyr::rename()}} method.
\subsection{\code{SpatRaster}}{

Rename layers of a \code{SpatRaster}.
}

\subsection{\code{SpatVector}}{

The result is a \code{SpatVector} with the renamed attributes on the function
call.
}
}

\examples{

library(terra)
f <- system.file("extdata/cyl_tile.tif", package = "tidyterra")
spatrast <- rast(f) |> mutate(aa = 1, bb = 2, cc = 3)

spatrast

spatrast |> rename(
  this_first = cyl_tile_1,
  this_second = cyl_tile_2
)

spatrast |> rename_with(
  toupper,
  .cols = starts_with("c")
)
}
\seealso{
\code{\link[dplyr:rename]{dplyr::rename()}}

Other single table verbs: 
\code{\link{arrange.SpatVector}()},
\code{\link{filter.Spat}},
\code{\link{mutate.Spat}},
\code{\link{select.Spat}},
\code{\link{slice.Spat}},
\code{\link{summarise.SpatVector}()}

Other \CRANpkg{dplyr} verbs that operate on columns:
\code{\link{glimpse.Spat}},
\code{\link{mutate.Spat}},
\code{\link{pull.Spat}},
\code{\link{relocate.Spat}},
\code{\link{select.Spat}}

Other \CRANpkg{dplyr} methods:
\code{\link{arrange.SpatVector}()},
\code{\link{bind_cols.SpatVector}},
\code{\link{bind_rows.SpatVector}},
\code{\link{count.SpatVector}()},
\code{\link{distinct.SpatVector}()},
\code{\link{filter-joins.SpatVector}},
\code{\link{filter.Spat}},
\code{\link{glimpse.Spat}},
\code{\link{group-by.SpatVector}},
\code{\link{mutate-joins.SpatVector}},
\code{\link{mutate.Spat}},
\code{\link{pull.Spat}},
\code{\link{relocate.Spat}},
\code{\link{rowwise.SpatVector}()},
\code{\link{select.Spat}},
\code{\link{slice.Spat}},
\code{\link{summarise.SpatVector}()}
}
\concept{dplyr.cols}
\concept{dplyr.methods}
\concept{single table verbs}
