% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows-SpatVector.R
\encoding{UTF-8}
\name{rows.SpatVector}
\alias{rows.SpatVector}
\alias{rows_insert.SpatVector}
\alias{rows_append.SpatVector}
\alias{rows_update.SpatVector}
\alias{rows_patch.SpatVector}
\alias{rows_upsert.SpatVector}
\alias{rows_delete.SpatVector}
\title{Row operations for \code{SpatVector} objects}
\usage{
\method{rows_insert}{SpatVector}(
  x,
  y,
  by = NULL,
  ...,
  conflict = c("error", "ignore"),
  copy = FALSE,
  in_place = FALSE
)

\method{rows_append}{SpatVector}(x, y, ..., copy = FALSE, in_place = FALSE)

\method{rows_update}{SpatVector}(
  x,
  y,
  by = NULL,
  ...,
  unmatched = c("error", "ignore"),
  copy = FALSE,
  in_place = FALSE
)

\method{rows_patch}{SpatVector}(
  x,
  y,
  by = NULL,
  ...,
  unmatched = c("error", "ignore"),
  copy = FALSE,
  in_place = FALSE
)

\method{rows_upsert}{SpatVector}(x, y, by = NULL, ..., copy = FALSE, in_place = FALSE)

\method{rows_delete}{SpatVector}(
  x,
  y,
  by = NULL,
  ...,
  unmatched = c("error", "ignore"),
  copy = FALSE,
  in_place = FALSE
)
}
\arguments{
\item{x}{A \code{SpatVector}.}

\item{y}{A data frame, \code{sf} object or \code{SpatVector}.}

\item{by}{An unnamed character vector giving the key columns. The key columns
must exist in both \code{x} and \code{y}. Keys typically uniquely identify each row,
but this is only enforced for the key values of \code{y} when \code{rows_update()},
\code{rows_patch()}, or \code{rows_upsert()} are used.

By default, we use the first column in \code{y}, since the first column is
a reasonable place to put an identifier variable.}

\item{...}{Other parameters passed onto methods.}

\item{conflict}{For \code{rows_insert()}, how should keys in \code{y} that conflict
with keys in \code{x} be handled? A conflict arises if there is a key in \code{y}
that already exists in \code{x}.

One of:
\itemize{
\item \code{"error"}, the default, will error if there are any keys in \code{y} that
conflict with keys in \code{x}.
\item \code{"ignore"} will ignore rows in \code{y} with keys that conflict with keys in
\code{x}.
}}

\item{copy}{If \code{x} and \code{y} are not from the same data source,
and \code{copy} is \code{TRUE}, then \code{y} will be copied into the
same src as \code{x}.  This allows you to join tables across srcs, but
it is a potentially expensive operation so you must opt into it.}

\item{in_place}{Should \code{x} be modified in place? This argument is only
relevant for mutable backends (e.g. databases, data.tables).

When \code{TRUE}, a modified version of \code{x} is returned invisibly;
when \code{FALSE}, a new object representing the resulting changes is returned.}

\item{unmatched}{For \code{rows_update()}, \code{rows_patch()}, and \code{rows_delete()},
how should keys in \code{y} that are unmatched by the keys in \code{x} be handled?

One of:
\itemize{
\item \code{"error"}, the default, will error if there are any keys in \code{y} that
are unmatched by the keys in \code{x}.
\item \code{"ignore"} will ignore rows in \code{y} with keys that are unmatched by the
keys in \code{x}.
}}
}
\value{
A \code{SpatVector}.
}
\description{
Methods for the \code{\link[dplyr:rows_insert]{dplyr::rows_insert()}} family on \code{SpatVector} objects.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:rows_insert]{dplyr::rows_insert()}} family for
\code{SpatVector} objects.
\subsection{\code{SpatVector}}{

Row operations update attributes while preserving the geometry column. When
inserting data frame rows without geometry, the output contains empty
geometries for the new rows.
}
}

\examples{
v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))

rows_update(
  v,
  tibble::tibble(cpro = "05", name = "New name"),
  by = "cpro"
)

rows_insert(
  v,
  tibble::tibble(cpro = "99", name = "New province"),
  by = "cpro"
)
}
\seealso{
\code{\link[dplyr:rows_insert]{dplyr::rows_insert()}}

Other \CRANpkg{dplyr} verbs that operate on rows:
\code{\link[=arrange.SpatVector]{arrange.SpatVector()}},
\code{\link[=distinct.SpatVector]{distinct.SpatVector()}},
\code{\link{filter.Spat}},
\code{\link{slice.Spat}}

Other \CRANpkg{dplyr} verbs that operate on pairs of \code{SpatVector} and data frame objects:
\code{\link{bind_cols.SpatVector}},
\code{\link{bind_rows.SpatVector}},
\code{\link[=cross_join.SpatVector]{cross_join.SpatVector()}},
\code{\link{filter-joins.SpatVector}},
\code{\link{mutate-joins.SpatVector}},
\code{\link[=nest_join.SpatVector]{nest_join.SpatVector()}}

Other \CRANpkg{dplyr} methods:
\code{\link[=arrange.SpatVector]{arrange.SpatVector()}},
\code{\link{bind_cols.SpatVector}},
\code{\link{bind_rows.SpatVector}},
\code{\link[=count.SpatVector]{count.SpatVector()}},
\code{\link[=cross_join.SpatVector]{cross_join.SpatVector()}},
\code{\link[=distinct.SpatVector]{distinct.SpatVector()}},
\code{\link{filter-joins.SpatVector}},
\code{\link{filter.Spat}},
\code{\link{glimpse.Spat}},
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}},
\code{\link{mutate-joins.SpatVector}},
\code{\link{mutate.Spat}},
\code{\link[=nest_join.SpatVector]{nest_join.SpatVector()}},
\code{\link{pull.Spat}},
\code{\link[=reframe.SpatVector]{reframe.SpatVector()}},
\code{\link{relocate.Spat}},
\code{\link{rename.Spat}},
\code{\link[=rowwise.SpatVector]{rowwise.SpatVector()}},
\code{\link{select.Spat}},
\code{\link{slice.Spat}},
\code{\link[=summarise.SpatVector]{summarise.SpatVector()}}
}
\concept{dplyr.methods}
\concept{dplyr.pairs}
\concept{dplyr.rows}
