% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise-SpatVector.R
\encoding{UTF-8}
\name{rowwise.SpatVector}
\alias{rowwise.SpatVector}
\title{Group \code{\link[terra:vect]{SpatVector}} objects by rows}
\usage{
\method{rowwise}{SpatVector}(data, ...)
}
\arguments{
\item{data}{A \code{SpatVector} object. See \strong{Methods}.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Variables to be
preserved when calling \code{\link[=summarise.SpatVector]{summarise.SpatVector()}}. This is typically a set
of variables whose combination uniquely identifies each row. See
\code{\link[dplyr:rowwise]{dplyr::rowwise()}}.

Unlike \code{\link[=group_by.SpatVector]{group_by.SpatVector()}}, you cannot create new variables here.
Instead, you can select multiple variables, for example with
\code{\link[tidyselect:everything]{everything()}}.}
}
\value{
A \code{\link[terra:vect]{SpatVector}} object with updated grouping metadata.
}
\description{
\code{rowwise()} lets you compute on a \code{\link[terra:vect]{SpatVector}} one row at a
time.
This is most useful when a vectorized function does not exist.

Most \CRANpkg{dplyr} verb implementations in \CRANpkg{tidyterra} preserve
row-wise grouping. The exception is \code{\link[=summarise.SpatVector]{summarise.SpatVector()}}, which returns
a \link[=group_by.SpatVector]{grouped \code{SpatVector}}. You can explicitly ungroup
with
\code{\link[=ungroup.SpatVector]{ungroup.SpatVector()}} or \code{\link[=as_tibble]{as_tibble()}} or convert to a grouped \code{SpatVector}
with \code{\link[=group_by.SpatVector]{group_by.SpatVector()}}.
}
\details{
See \strong{Details} on \code{\link[dplyr:rowwise]{dplyr::rowwise()}}.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:rowwise]{dplyr::rowwise()}} method for
\code{\link[terra:vect]{SpatVector}} objects.
}

\section{Grouping metadata}{


Mixing \CRANpkg{terra} and \CRANpkg{dplyr} syntax on a grouped or row-wise
\code{\link[terra:vect]{SpatVector}}, for example by subsetting with \code{v[1:3, 1:2]},
can corrupt its
grouping metadata. \CRANpkg{tidyterra} attempts to restore this metadata the
next time you use a \CRANpkg{dplyr} verb on the object.

Some operations, such as \code{\link[terra:spatSample]{terra::spatSample()}}, create a new
\code{\link[terra:vect]{SpatVector}}
without preserving grouping metadata. Call \code{\link[=group_by.SpatVector]{group_by.SpatVector()}} or
\code{\link[=rowwise.SpatVector]{rowwise.SpatVector()}} again, as appropriate.
}

\examples{
library(terra)
library(dplyr)

v <- terra::vect(system.file("shape/nc.shp", package = "sf"))

# Select new births
nb <- v |>
  select(starts_with("NWBIR")) |>
  glimpse()

# Compute the mean of NWBIR for each geometry.
nb |>
  rowwise() |>
  mutate(nb_mean = mean(c(NWBIR74, NWBIR79)))

# Additional examples
\donttest{
# Use c_across() to select many variables more easily.
nb |>
  rowwise() |>
  mutate(m = mean(c_across(NWBIR74:NWBIR79)))

# Compute the minimum of x and y in each row

nb |>
  rowwise() |>
  mutate(min = min(c_across(NWBIR74:NWBIR79)))

# Summarize.
v |>
  rowwise() |>
  summarise(mean_bir = mean(BIR74, BIR79)) |>
  glimpse() |>
  autoplot(aes(fill = mean_bir))

# Supply a variable to be kept
v |>
  mutate(id2 = as.integer(CNTY_ID / 100)) |>
  rowwise(id2) |>
  summarise(mean_bir = mean(BIR74, BIR79)) |>
  glimpse() |>
  autoplot(aes(fill = as.factor(id2)))
}
}
\seealso{
\code{\link[dplyr:rowwise]{dplyr::rowwise()}}.

Other \CRANpkg{dplyr} verbs that operate on groups of rows:
\code{\link[=count.SpatVector]{count.SpatVector()}},
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}},
\code{\link[=reframe.SpatVector]{reframe.SpatVector()}},
\code{\link[=summarise.SpatVector]{summarise.SpatVector()}}
}
\concept{dplyr.groups}
