% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise-SpatVector.R
\encoding{UTF-8}
\name{rowwise.SpatVector}
\alias{rowwise.SpatVector}
\title{Group \code{SpatVector} objects by rows}
\usage{
\method{rowwise}{SpatVector}(data, ...)
}
\arguments{
\item{data}{A \code{SpatVector} object. See \strong{Methods}.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Variables to be
preserved when calling \code{\link[=summarise.SpatVector]{summarise.SpatVector()}}. This is typically a set of
variables whose combination uniquely identifies each row. See
\code{\link[dplyr:rowwise]{dplyr::rowwise()}}.

Unlike \code{\link[=group_by.SpatVector]{group_by.SpatVector()}}, you cannot create new variables here.
Instead, you can select multiple variables, for example with
\code{\link[tidyselect:everything]{everything()}}.}
}
\value{
The same \code{SpatVector} object with updated grouping metadata.
}
\description{
\code{rowwise()} lets you compute on a \code{SpatVector} one row at a time.
This is most useful when a vectorised function does not exist.

Most \CRANpkg{dplyr} verb implementations in \CRANpkg{tidyterra} preserve
row-wise grouping. The exception is \code{\link[=summarise.SpatVector]{summarise.SpatVector()}}, which returns
a \link[=group_by.SpatVector]{grouped SpatVector}. You can explicitly ungroup with
\code{\link[=ungroup.SpatVector]{ungroup.SpatVector()}} or \code{\link[=as_tibble]{as_tibble()}} or convert to a grouped \code{SpatVector}
with \code{\link[=group_by.SpatVector]{group_by.SpatVector()}}.
}
\details{
See \strong{Details} on \code{\link[dplyr:rowwise]{dplyr::rowwise()}}.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[dplyr:rowwise]{dplyr::rowwise()}} function for
\code{SpatVector} objects.

\strong{When mixing} \CRANpkg{terra} \strong{and} \CRANpkg{dplyr} \strong{syntax} on a
row-wise \code{SpatVector}, for example subsetting a \code{SpatVector} like
\code{v[1:3,1:2]}, the \code{groups} attribute can be corrupted.
\CRANpkg{tidyterra} tries to
regenerate the \code{SpatVector}. This is triggered the next time you use a
\CRANpkg{dplyr} verb on your \code{SpatVector}.

Some operations, such as \code{terra::spatSample()}, create a new \code{SpatVector}.
In these cases, the result does not preserve the \code{groups} attribute. Use
\code{\link[=rowwise.SpatVector]{rowwise.SpatVector()}} to re-group.
}

\examples{
library(terra)
library(dplyr)

v <- terra::vect(system.file("shape/nc.shp", package = "sf"))

# Select new births
nb <- v |>
  select(starts_with("NWBIR")) |>
  glimpse()

# Compute the mean of NWBIR on each geometry
nb |>
  rowwise() |>
  mutate(nb_mean = mean(c(NWBIR74, NWBIR79)))

# Additional examples
\donttest{
# Use c_across() to select many variables more easily.
nb |>
  rowwise() |>
  mutate(m = mean(c_across(NWBIR74:NWBIR79)))

# Compute the minimum of x and y in each row

nb |>
  rowwise() |>
  mutate(min = min(c_across(NWBIR74:NWBIR79)))

# Summarize.
v |>
  rowwise() |>
  summarise(mean_bir = mean(BIR74, BIR79)) |>
  glimpse() |>
  autoplot(aes(fill = mean_bir))

# Supply a variable to be kept
v |>
  mutate(id2 = as.integer(CNTY_ID / 100)) |>
  rowwise(id2) |>
  summarise(mean_bir = mean(BIR74, BIR79)) |>
  glimpse() |>
  autoplot(aes(fill = as.factor(id2)))
}
}
\seealso{
\code{\link[dplyr:rowwise]{dplyr::rowwise()}}

Other \CRANpkg{dplyr} verbs that operate on groups of rows:
\code{\link[=count.SpatVector]{count.SpatVector()}},
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}},
\code{\link[=reframe.SpatVector]{reframe.SpatVector()}},
\code{\link[=summarise.SpatVector]{summarise.SpatVector()}}

Other \CRANpkg{dplyr} methods:
\code{\link[=arrange.SpatVector]{arrange.SpatVector()}},
\code{\link{bind_cols.SpatVector}},
\code{\link{bind_rows.SpatVector}},
\code{\link[=count.SpatVector]{count.SpatVector()}},
\code{\link[=cross_join.SpatVector]{cross_join.SpatVector()}},
\code{\link[=distinct.SpatVector]{distinct.SpatVector()}},
\code{\link{filter-joins.SpatVector}},
\code{\link{filter.Spat}},
\code{\link{glimpse.Spat}},
\code{\link[=group_by.SpatVector]{group_by.SpatVector()}},
\code{\link{mutate-joins.SpatVector}},
\code{\link{mutate.Spat}},
\code{\link[=nest_join.SpatVector]{nest_join.SpatVector()}},
\code{\link{pull.Spat}},
\code{\link[=reframe.SpatVector]{reframe.SpatVector()}},
\code{\link{relocate.Spat}},
\code{\link{rename.Spat}},
\code{\link{rows.SpatVector}},
\code{\link{select.Spat}},
\code{\link{slice.Spat}},
\code{\link[=summarise.SpatVector]{summarise.SpatVector()}}
}
\concept{dplyr.groups}
\concept{dplyr.methods}
