% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyterra-package.R
\docType{package}
\encoding{UTF-8}
\name{tidyterra-package}
\alias{tidyterra}
\alias{tidyterra-package}
\title{tidyterra: 'tidyverse' Methods and 'ggplot2' Helpers for 'terra' Objects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides methods from 'tidyverse' packages for 'SpatRaster' and 'SpatVector' objects created with 'terra', plus 'ggplot2' 'geoms' and scales for plotting those objects. It is designed to let users manipulate spatial data with familiar 'dplyr' and 'tidyr' verbs before visualizing results with 'ggplot2'.
}
\references{
Hernangomez D (2023). "Using the tidyverse with terra objects: the
tidyterra package." \emph{Journal of Open Source Software},
\strong{8}(91), 5751. ISSN 2475-9066. \doi{10.21105/joss.05751}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/tidyterra/}
  \item \url{https://github.com/dieghernan/tidyterra}
  \item Report bugs at \url{https://github.com/dieghernan/tidyterra/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Dewey Dunnington (\href{https://orcid.org/0000-0002-9415-4582}{ORCID}) (for ggspatial code) [contributor]
  \item ggplot2 authors (for contour code) [copyright holder]
  \item Andrea Manica (\href{https://orcid.org/0000-0003-1895-450X}{ORCID}) [contributor]
}

}
\keyword{internal}
