% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyterra-package.R
\docType{package}
\name{tidyterra-package}
\alias{tidyterra}
\alias{tidyterra-package}
\title{tidyterra: 'tidyverse' Methods for 'terra' Objects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extension of the 'tidyverse' for 'SpatRaster' and 'SpatVector' objects of the 'terra' package. Includes also new 'geom_' functions that provide a convenient way of visualizing 'terra' objects with 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/tidyterra/}
  \item \url{https://github.com/dieghernan/tidyterra}
  \item Report bugs at \url{https://github.com/dieghernan/tidyterra/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Dewey Dunnington (\href{https://orcid.org/0000-0002-9415-4582}{ORCID}) (for ggspatial code) [contributor]
}

}
\keyword{internal}
