% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmute-Spat.R
\name{transmute.Spat}
\alias{transmute.Spat}
\alias{transmute.SpatRaster}
\alias{transmute.SpatVector}
\title{Create, modify, and delete cell values/layers/attributes of \verb{Spat*} objects}
\usage{
\method{transmute}{SpatRaster}(.data, ...)

\method{transmute}{SpatVector}(.data, ...)
}
\arguments{
\item{.data}{A \code{SpatRaster} created with \code{\link[terra:rast]{terra::rast()}} or a \code{SpatVector}
created with \code{\link[terra:vect]{terra::vect()}}.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Name-value pairs.
The name gives the name of the column in the output.

The value can be:
\itemize{
\item A vector of length 1, which will be recycled to the correct length.
\item A vector the same length as the current group (or the whole data frame
if ungrouped).
\item \code{NULL}, to remove the column.
\item A data frame or tibble, to create multiple columns in the output.
}}
}
\value{
A \verb{Spat*} object of the same class than \code{.data}. See \strong{Methods}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

\code{transmute()} creates a new object containing only the specified
computations. It's superseded because you can perform the same job
with \code{mutate(.keep = "none")}.
}
\section{Methods}{

Implementation of the \strong{generic} \code{\link[dplyr:transmute]{dplyr::transmute()}} method.
}

\examples{
library(terra)

# SpatVector method
f <- system.file("extdata/cyl.gpkg", package = "tidyterra")
v <- vect(f)

v |>
  transmute(cpro2 = paste0(cpro, "-CyL"))
}
\seealso{
\code{\link{mutate.Spat}}, \code{\link[dplyr:transmute]{dplyr::transmute()}} methods.
}
\keyword{internal}
