% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncount-SpatVector.R
\encoding{UTF-8}
\name{uncount.SpatVector}
\alias{uncount.SpatVector}
\title{Duplicate \code{SpatVector} rows}
\usage{
\method{uncount}{SpatVector}(data, weights, ..., .remove = TRUE, .id = NULL)
}
\arguments{
\item{data}{A \code{SpatVector}.}

\item{weights}{A vector of weights. Evaluated in the context of \code{data};
supports quasiquotation.}

\item{...}{Additional arguments passed on to methods.}

\item{.remove}{If \code{TRUE}, and \code{weights} is the name of a column in \code{data},
then this column is removed.}

\item{.id}{Supply a string to create a new variable which gives a unique
identifier for each created row.}
}
\value{
A \code{SpatVector} object.
}
\description{
\code{uncount()} duplicates rows according to a weighting variable.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[tidyr:uncount]{tidyr::uncount()}} method.
\subsection{\code{SpatVector}}{

Each duplicated row keeps the input geometry.
}
}

\examples{
library(tidyr)

v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))
v$copies <- rep_len(1:2, nrow(v))

uncount(v, copies)

}
\seealso{
\code{\link[tidyr:uncount]{tidyr::uncount()}}

Other \CRANpkg{tidyr} methods:
\code{\link[=complete.SpatVector]{complete.SpatVector()}},
\code{\link{drop_na.Spat}},
\code{\link[=expand.SpatVector]{expand.SpatVector()}},
\code{\link[=fill.SpatVector]{fill.SpatVector()}},
\code{\link[=nest.SpatVector]{nest.SpatVector()}},
\code{\link[=pivot_longer.SpatVector]{pivot_longer.SpatVector()}},
\code{\link[=pivot_wider.SpatVector]{pivot_wider.SpatVector()}},
\code{\link{replace_na.Spat}},
\code{\link{unite.Spat}}
}
\concept{tidyr.methods}
\concept{tidyr.rows}
