% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite-Spat.R
\encoding{UTF-8}
\name{unite.Spat}
\alias{unite.Spat}
\alias{unite.SpatRaster}
\alias{unite.SpatVector}
\title{Unite \verb{Spat*} layers or attributes}
\usage{
\method{unite}{SpatRaster}(data, col, ..., sep = "_", remove = TRUE, na.rm = FALSE)

\method{unite}{SpatVector}(data, col, ..., sep = "_", remove = TRUE, na.rm = FALSE)
}
\arguments{
\item{data}{A \code{SpatRaster} or \code{SpatVector}.}

\item{col}{The name of the new column, as a string or symbol.

This argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote strings
and symbols). The name is captured from the expression with
\code{\link[rlang:defusing-advanced]{rlang::ensym()}} (note that this kind of interface where
symbols do not represent actual objects is now discouraged in the
tidyverse; we support it here for backward compatibility).}

\item{...}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> Columns to unite}

\item{sep}{Separator to use between values.}

\item{remove}{If \code{TRUE}, remove input columns from output data frame.}

\item{na.rm}{If \code{TRUE}, missing values will be removed prior to uniting
each value.}
}
\value{
A \code{SpatRaster} or \code{SpatVector} object.
}
\description{
\code{unite()} combines multiple layers or attributes by pasting their values
together.
}
\section{Methods}{


Implementation of the \strong{generic} \code{\link[tidyr:unite]{tidyr::unite()}} method.
\subsection{\code{SpatRaster}}{

The selected layers are united cell by cell. The new layer is categorical
because \code{\link[tidyr:unite]{tidyr::unite()}} returns a character vector.
}

\subsection{\code{SpatVector}}{

The geometry column has sticky behavior and is never united with attributes.
}
}

\examples{
library(tidyr)

v <- terra::vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))

unite(v, "label", iso2, cpro, sep = "-")

r <- terra::rast(system.file("extdata/cyl_temp.tif", package = "tidyterra"))

unite(r, "label", tavg_04, tavg_05, sep = "-", remove = FALSE)

}
\seealso{
\code{\link[tidyr:unite]{tidyr::unite()}}

Other \CRANpkg{tidyr} methods:
\code{\link[=complete.SpatVector]{complete.SpatVector()}},
\code{\link{drop_na.Spat}},
\code{\link[=expand.SpatVector]{expand.SpatVector()}},
\code{\link[=fill.SpatVector]{fill.SpatVector()}},
\code{\link[=nest.SpatVector]{nest.SpatVector()}},
\code{\link[=pivot_longer.SpatVector]{pivot_longer.SpatVector()}},
\code{\link[=pivot_wider.SpatVector]{pivot_wider.SpatVector()}},
\code{\link{replace_na.Spat}},
\code{\link[=uncount.SpatVector]{uncount.SpatVector()}}
}
\concept{tidyr.character}
\concept{tidyr.methods}
