% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{volcano2}
\alias{volcano2}
\title{Updated Topographic Information on Auckland's Maungawhau Volcano}
\format{
A matrix of 109 rows and 76 columns.
}
\source{
\href{https://data.linz.govt.nz/}{Toitu Te Whenua Land Information New Zealand}

8m Digital Elevation Model (DEM) created by Geographx
(\url{https://geographx.co.nz/}), primarily derived from January 2012 LINZ
Topo50 20m contours (data.linz.govt.nz/layer/768).
}
\description{
Updated Topographic Information on Auckland's Maungawhau Volcano
}
\seealso{
\link{volcano}
}
\concept{datasets}
